@@ -18,69 +18,74 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.17.6 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  sqflite: any
-  path_provider: any
-  shared_preferences: any
+  
 
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
+  
+  path_provider: any
+  shared_preferences: any
+
   cupertino_icons: ^1.0.0
-  font_awesome_flutter: ^8.8.1
-  smooth_star_rating: ^1.1.1
-  intl: ^0.16.0
-  http: ^0.12.0+4
-  geolocator: ^6.1.6
-  sprintf: "^4.0.0"
-  connectivity: ^0.4.2
-  flutter_html: ^0.8.2
-  fluttertoast: ^4.0.0
-  get_it: ^5.0.1
-  carousel_slider: ^2.3.1
-  google_maps_flutter: ^1.0.2
-  animated_stack: ^0.1.0
-  url_launcher: ^5.0.3
-  provider: ^4.0.2
-  sliding_up_panel: ^1.0.2
-  flutter_spinkit: "^4.1.2"
-  rounded_loading_button: ^1.0.0
-  flutter_login_facebook: ^0.3.0
-  flutter_twitter: ^1.1.3
-  crypto: ^2.1.5
+  font_awesome_flutter: ^10.1.0
+  smooth_star_rating_nsafe: ^1.0.0+1
+  intl: ^0.17.0
+  http: ^0.13.5
+  geolocator: ^9.0.1
+  sprintf: ^6.0.2
+ 
+  flutter_html: ^2.2.1
+  fluttertoast: ^8.0.9
+  get_it: ^7.2.0
+  carousel_slider: ^4.1.1
+  google_maps_flutter: ^2.1.12
+  animated_stack: ^0.2.0
+  url_launcher: ^6.1.5
+  provider: ^6.0.3
+  sliding_up_panel: ^2.0.0+1
+  flutter_spinkit: ^5.1.0
+  rounded_loading_button: ^2.1.0
+  flutter_login_facebook: ^1.6.0
+  package_info_plus: ^1.4.1
+  twitter_login: ^4.2.3
+  crypto: ^3.0.2
   random_string: ^2.1.0
-  image_picker: ^0.6.7+12
+  image_picker: ^0.8.5+3
   flutter_custom_dialog: ^1.0.20
-  ndialog: ^3.1.0
-  direct_select: "^1.0.3"
-  day_night_time_picker: ^0.3.0+2
-  cached_network_image: ^2.3.3
-  social_share_plugin: ^0.3.1+1
-  onesignal_flutter: ^2.0.0
-  string_validator: ^0.1.4
+  ndialog: ^4.1.1+1
+  direct_select: ^2.0.0
+  day_night_time_picker: ^1.1.2
+  cached_network_image: ^3.2.1
+  social_share_plugin: ^0.4.1+2
+  onesignal_flutter: ^3.4.1
+  string_validator: ^0.3.0
 
 
   # flutter_zoom_drawer:
   #   git:
   #     url: https://github.com/mgflutter/flutter_zoom_drawer
 
-  easy_localization: ^2.3.3
-  incrementally_loading_listview: ^0.3.0
-  shimmer: ^1.1.2
-  html_unescape: ^1.0.2
+  easy_localization: ^3.0.1
+  incrementally_loading_listview: ^1.0.0
+  shimmer: ^2.0.0
+  html_unescape: ^2.0.0
   image_viewer: ^1.0.0
-  photo_view: ^0.10.2
-  drag_down_to_pop: ^0.2.1
-  get_version: ^0.2.2
+  photo_view: ^0.14.0
+  drag_down_to_pop: ^1.0.0
+  #get_version: ^0.2.2
   # android_intent: ^0.3.6+12
   # permission_handler: ^5.0.1+1
   # location_permissions: ^3.0.0
-  flutter_cache_manager: ^2.0.0
+  flutter_cache_manager: ^3.3.0
+  sqflite: ^2.0.3+1
+  connectivity_plus: ^2.3.6+1
 
 
 dev_dependencies:
