@@ -6,13 +6,13 @@ class AppBuilder extends StatefulWidget {
   final Function(BuildContext) builder;
 
   const AppBuilder(
-      {Key key, this.builder})
+      {Key? key, required this.builder})
       : super(key: key);
 
   @override
   AppBuilderState createState() => new AppBuilderState();
 
-  static AppBuilderState of(BuildContext context) {
+  static AppBuilderState? of(BuildContext context) {
     // return context.ancestorStateOfType(const TypeMatcher<AppBuilderState>());
     return context.findAncestorStateOfType<AppBuilderState>();
   }
