@@ -13,23 +13,22 @@ import 'package:directorio_delicias/commons/services.dart';
 import 'package:directorio_delicias/notifiers/datahandler_notifier.dart';
 
 class AboutUsPage extends StatefulWidget {
-
-  const AboutUsPage({Key key}) : super(key: key);
+  const AboutUsPage({Key? key}) : super(key: key);
 
   @override
   _AboutUsPageState createState() => _AboutUsPageState();
 }
 
-class _AboutUsPageState extends State<AboutUsPage> with TickerProviderStateMixin {
-
+class _AboutUsPageState extends State<AboutUsPage>
+    with TickerProviderStateMixin {
   final double infoHeight = 364.0;
-  AnimationController animationController;
-  Animation<double> animation;
+  late AnimationController animationController;
+  late Animation<double> animation;
   double opacity1 = 0.0;
   double opacity2 = 0.0;
   double opacity3 = 0.0;
   double opacity4 = 0.0;
-  
+
   @override
   void initState() {
     animationController = AnimationController(
@@ -54,135 +53,135 @@ class _AboutUsPageState extends State<AboutUsPage> with TickerProviderStateMixin
   @override
   Widget build(BuildContext context) {
     return ChangeNotifierProvider<DataHandlerNotifier>(
-      create: (context) => DataHandlerNotifier(),
-      builder: (ctx, widget) {
-        return root();
-      }
-    );
+        create: (context) => DataHandlerNotifier(),
+        builder: (ctx, widget) {
+          return root();
+        });
   }
-  
+
   Future<bool> getData() async {
     await Future<dynamic>.delayed(const Duration(milliseconds: 200));
     return true;
   }
 
   Widget root() {
-
     var left;
     var right;
 
-    if(Helpers.isRTL()) {
+    if (Helpers.isRTL()) {
       left = 35.0;
-    }
-    else {
+    } else {
       right = 35.0;
     }
 
     return Container(
-      child: Column(
+        child: Column(children: <Widget>[
+      Stack(
         children: <Widget>[
-          Stack(
+          Column(
             children: <Widget>[
-              Column(
-                children: <Widget>[
-                  Container(
-                    color:Theme.of(context).accentColor,
-                    height: MediaQuery.of(context).size.height,
-                  ),
+              Container(
+                color: Theme.of(context).accentColor,
+                height: MediaQuery.of(context).size.height,
+              ),
+            ],
+          ),
+          Positioned(
+            top: (MediaQuery.of(context).size.width / 3.0) - 24.0,
+            bottom: 0,
+            left: 0,
+            right: 0,
+            child: Container(
+              decoration: BoxDecoration(
+                color: Theme.of(context).backgroundColor,
+                borderRadius: const BorderRadius.only(
+                    topLeft: Radius.circular(32.0),
+                    topRight: Radius.circular(32.0)),
+                boxShadow: <BoxShadow>[
+                  BoxShadow(
+                      color: Theme.of(context).shadowColor,
+                      offset: const Offset(1.1, 1.1),
+                      blurRadius: 10.0),
                 ],
               ),
-              Positioned(
-                top: (MediaQuery.of(context).size.width / 3.0) - 24.0,
-                bottom: 0,
-                left: 0,
-                right: 0,
-                child: Container(
-                  decoration: BoxDecoration(
-                    color:Theme.of(context).backgroundColor,
-                    borderRadius: const BorderRadius.only(
-                        topLeft: Radius.circular(32.0),
-                        topRight: Radius.circular(32.0)),
-                    boxShadow: <BoxShadow>[
-                      BoxShadow(
-                          color:Theme.of(context).shadowColor,
-                          offset: const Offset(1.1, 1.1),
-                          blurRadius: 10.0),
-                    ],
-                  ),
-                  child: Padding(
-                    padding: const EdgeInsets.only(left: 8, right: 8, top: 40),
-                    child: SingleChildScrollView(
-                      child: Container(
-                        child: Column(
-                          mainAxisAlignment: MainAxisAlignment.center,
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: <Widget>[
-                            Padding(
-                                padding: const EdgeInsets.only(left: 16, right: 16, top: 8, bottom: 8),
-                                child: Html(
-                                  data: tr("about_us_details"),
-                                  defaultTextStyle: TextStyle(
-                                    fontWeight: FontWeight.w200,
-                                    fontSize: 14,
-                                    letterSpacing: 0.27,
-                                    color: Theme.of(context).textTheme.caption.color,
-                                  ),
+              child: Padding(
+                padding: const EdgeInsets.only(left: 8, right: 8, top: 40),
+                child: SingleChildScrollView(
+                  child: Container(
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: <Widget>[
+                        Padding(
+                            padding: const EdgeInsets.only(
+                                left: 16, right: 16, top: 8, bottom: 8),
+                            child: Html(
+                              data: tr("about_us_details"),
+                              style: {
+                                "*": Style(
+                                  fontWeight: FontWeight.w200,
+                                  fontSize: FontSize.em(14),
+                                  letterSpacing: 0.27,
+                                  color: Theme.of(context)
+                                      .textTheme
+                                      .caption!
+                                      .color,
                                 )
-                              ),
-                            
-                            SizedBox(height: MediaQuery.of(context).padding.bottom )
-                          ],
-                        ),
-                      ),
+                              },
+                            )),
+                        SizedBox(height: MediaQuery.of(context).padding.bottom)
+                      ],
                     ),
                   ),
                 ),
               ),
-              Positioned(
-                top: (MediaQuery.of(context).size.width / 3.0) - 24.0 - 35,
-                right: right,
-                left: left,
-                child: ScaleTransition(
-                  alignment: Alignment.center,
-                  scale: CurvedAnimation(
-                      parent: animationController, curve: Curves.fastOutSlowIn),
-                  child: Card(
-                    color:Theme.of(context).accentColor,
-                    shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(50.0)),
-                    elevation: 10.0,
-                    shadowColor: Theme.of(context).shadowColor,
-                    child: Consumer<DataHandlerNotifier>(
-                      builder: (context, provider, child) => InkWell(
-                        borderRadius: BorderRadius.circular(50.0),
-                        child: Container(
-                          width: 60,
-                          height: 60,
-                          child: Center(
-                            child: Icon(
-                              Icons.email,
-                              color: Theme.of(context).indicatorColor,
-                              size: 30,
-                            ),
-                          ),
+            ),
+          ),
+          Positioned(
+            top: (MediaQuery.of(context).size.width / 3.0) - 24.0 - 35,
+            right: right,
+            left: left,
+            child: ScaleTransition(
+              alignment: Alignment.center,
+              scale: CurvedAnimation(
+                  parent: animationController, curve: Curves.fastOutSlowIn),
+              child: Card(
+                color: Theme.of(context).accentColor,
+                shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(50.0)),
+                elevation: 10.0,
+                shadowColor: Theme.of(context).shadowColor,
+                child: Consumer<DataHandlerNotifier>(
+                  builder: (context, provider, child) => InkWell(
+                    borderRadius: BorderRadius.circular(50.0),
+                    child: Container(
+                      width: 60,
+                      height: 60,
+                      child: Center(
+                        child: Icon(
+                          Icons.email,
+                          color: Theme.of(context).indicatorColor,
+                          size: 30,
                         ),
-                        onTap: () {
-                          GetIt.instance.get<Services>().sendEmail(Config.CONTACT_EMAIL);
-                        },
                       ),
                     ),
+                    onTap: () {
+                      GetIt.instance
+                          .get<Services>()
+                          .sendEmail(Config.CONTACT_EMAIL);
+                    },
                   ),
                 ),
               ),
-              getAppBarUI(),
-            ],
-          )
-        ]
+            ),
+          ),
+          getAppBarUI(),
+        ],
       )
-    );
+    ]));
   }
 
-Widget getAppBarUI() {
+  Widget getAppBarUI() {
     return Container(
       decoration: BoxDecoration(
         color: Colors.transparent,
@@ -233,9 +232,7 @@ Widget getAppBarUI() {
               child: Row(
                 crossAxisAlignment: CrossAxisAlignment.center,
                 mainAxisAlignment: MainAxisAlignment.end,
-                children: <Widget>[
-                  
-                ],
+                children: <Widget>[],
               ),
             )
           ],
@@ -244,4 +241,3 @@ Widget getAppBarUI() {
     );
   }
 }
-
