@@ -1,4 +1,3 @@
-
 import 'package:drag_down_to_pop/drag_down_to_pop.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -9,20 +8,18 @@ import 'package:directorio_delicias/commons/helpers.dart';
 import 'package:directorio_delicias/models/photo.dart';
 
 class ImageViewerScreen extends StatefulWidget {
+  ImageViewerScreen({Key? key, this.photos}) : super(key: key);
 
-  ImageViewerScreen({Key key, this.photos}) : super(key: key);
-
-  final List<Photo> photos;
+  final List<Photo>? photos;
 
   @override
   _ImageViewerScreenState createState() => _ImageViewerScreenState();
 }
 
 class _ImageViewerScreenState extends State<ImageViewerScreen> {
-
-  List<String> imageUrls = List<String>();
+  List<String> imageUrls = <String>[];
   PhotoViewScaleStateController _controller = PhotoViewScaleStateController();
-  
+
   @override
   void initState() {
     super.initState();
@@ -30,13 +27,11 @@ class _ImageViewerScreenState extends State<ImageViewerScreen> {
 
   @override
   Widget build(BuildContext context) {
-
     var left;
     var right;
-    if(Helpers.isRTL()) {
+    if (Helpers.isRTL()) {
       right = 20.0;
-    }
-    else {
+    } else {
       left = 20.0;
     }
 
@@ -46,78 +41,75 @@ class _ImageViewerScreenState extends State<ImageViewerScreen> {
       },
       child: Material(
         child: Stack(
-        children: [
-          Container(
-            color: Colors.black,
-            child: PhotoViewGallery.builder(
-              scrollPhysics: const BouncingScrollPhysics(),
-              builder: (BuildContext context, int index) {
-                return PhotoViewGalleryPageOptions(
-                  maxScale: 4.0,
-                  minScale: PhotoViewComputedScale.contained,
-                  tightMode: true,
-                  imageProvider: NetworkImage(widget.photos[index].photoUrl),
-                  initialScale: PhotoViewComputedScale.contained,
-                  scaleStateController: _controller..scaleState = PhotoViewScaleState.initial,
-                );
-              },
-              itemCount: widget.photos.length,
-              loadFailedChild: Center(
-                child: Container(
-                  child: Icon(
-                    Icons.image_not_supported_outlined,
-                    size: 80,
-                    color: Colors.grey[600]
-                  )
-                ),
-              ),
-              loadingBuilder: (context, event) => Center(
-                child: Container(
-                  color: Colors.transparent,
-                  width: 20.0,
-                  height: 20.0,
-                  child: SpinKitWave(
-                    itemCount: 3,
-                    color: Theme.of(context).accentColor,
-                    size: 20,
+          children: [
+            Container(
+                color: Colors.black,
+                child: PhotoViewGallery.builder(
+                  scrollPhysics: const BouncingScrollPhysics(),
+                  builder: (BuildContext context, int index) {
+                    return PhotoViewGalleryPageOptions(
+                      maxScale: 4.0,
+                      minScale: PhotoViewComputedScale.contained,
+                      tightMode: true,
+                      imageProvider:
+                          NetworkImage(widget.photos![index].photoUrl),
+                      initialScale: PhotoViewComputedScale.contained,
+                      scaleStateController: _controller
+                        ..scaleState = PhotoViewScaleState.initial,
+                    );
+                  },
+                  itemCount: widget.photos!.length,
+
+                  // loadFailedChild: Center(
+                  //   child: Container(
+                  //       child: Icon(Icons.image_not_supported_outlined,
+                  //           size: 80, color: Colors.grey[600])),
+                  // ),
+                  loadingBuilder: (context, event) => Center(
+                    child: Container(
+                      color: Colors.transparent,
+                      width: 20.0,
+                      height: 20.0,
+                      child: SpinKitWave(
+                        itemCount: 3,
+                        color: Theme.of(context).accentColor,
+                        size: 20,
+                      ),
+                    ),
                   ),
-                ),
-              ),
-            )
-          ),
-          Positioned(
-            top: AppBar().preferredSize.height,
-            left: left,
-            right: right,
-            child: Material(
-              color: Colors.transparent,
-              child: InkWell(
-                borderRadius: const BorderRadius.all(
-                  Radius.circular(50.0),
-                ),
-                child: Padding(
+                )),
+            Positioned(
+              top: AppBar().preferredSize.height,
+              left: left,
+              right: right,
+              child: Material(
+                color: Colors.transparent,
+                child: InkWell(
+                  borderRadius: const BorderRadius.all(
+                    Radius.circular(50.0),
+                  ),
+                  child: Padding(
                     padding: const EdgeInsets.all(8.0),
                     child: Icon(
                       Icons.close,
                       color: Colors.white,
                     ),
+                  ),
+                  onTap: () {
+                    Navigator.maybePop(context);
+                  },
                 ),
-                onTap: () {
-                  Navigator.maybePop(context);
-                },
               ),
-            ),
-          )
-        ],
-      ),
+            )
+          ],
+        ),
       ),
     );
-
   }
 }
 
 class ImageViewerPageRoute extends MaterialPageRoute {
-  ImageViewerPageRoute({@required WidgetBuilder builder})
+  ImageViewerPageRoute({required WidgetBuilder builder})
       : super(builder: builder, maintainState: false);
 
   @override
