@@ -1,4 +1,3 @@
-
 import 'dart:async';
 import 'dart:io';
 import 'package:easy_localization/easy_localization.dart';
@@ -11,52 +10,46 @@ import 'package:directorio_delicias/commons/helpers.dart';
 import 'package:directorio_delicias/dataparser/data_handler.dart';
 import 'package:directorio_delicias/fetcher/auth_fetcher.dart';
 
-
 class RegisterScreen extends StatefulWidget {
-
-  RegisterScreen({Key key}) : super(key: key);
+  RegisterScreen({Key? key}) : super(key: key);
 
   @override
   _RegisterScreenState createState() => _RegisterScreenState();
 }
 
 class _RegisterScreenState extends State<RegisterScreen> {
-  
   @override
   void initState() {
     super.initState();
   }
 
-  final RoundedLoadingButtonController _btnController = new RoundedLoadingButtonController();
+  final RoundedLoadingButtonController _btnController =
+      new RoundedLoadingButtonController();
   final TextEditingController _usernameController = new TextEditingController();
   final TextEditingController _passwordController = new TextEditingController();
   final TextEditingController _fullNameController = new TextEditingController();
   final TextEditingController _emailController = new TextEditingController();
   var photoFile;
   final picker = ImagePicker();
-  
+
   void register() async {
-    AuthFetcher(
-      didError: (DataHandler dataHandler) {
-        _btnController.error();
-        Timer(Duration(seconds: 2), () {
-          _btnController.reset();
-        });
-      },
-      didLogged: (DataHandler dataHandler) {
-        _btnController.success();
-        Timer(Duration(seconds: 2), () {
-          Navigator.pop(context, dataHandler);
-        });
-      }
-    ).register(
-      username: _usernameController.text, 
-      password: _passwordController.text, 
-      email: _emailController.text, 
-      fullName: _fullNameController.text, 
-      photoFile: photoFile != null ? photoFile.path : null
-    );
-  } 
+    AuthFetcher(didError: (DataHandler dataHandler) {
+      _btnController.error();
+      Timer(Duration(seconds: 2), () {
+        _btnController.reset();
+      });
+    }, didLogged: (DataHandler dataHandler) {
+      _btnController.success();
+      Timer(Duration(seconds: 2), () {
+        Navigator.pop(context, dataHandler);
+      });
+    }).register(
+        username: _usernameController.text,
+        password: _passwordController.text,
+        email: _emailController.text,
+        fullName: _fullNameController.text,
+        photoFile: photoFile != null ? photoFile.path : null);
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -67,13 +60,12 @@ class _RegisterScreenState extends State<RegisterScreen> {
         child: Scaffold(
           backgroundColor: Theme.of(context).backgroundColor,
           body: Column(
-            mainAxisAlignment: MainAxisAlignment.start,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: <Widget>[
-              getAppBarUI(),
-              main(),
-            ]
-          ),
+              mainAxisAlignment: MainAxisAlignment.start,
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: <Widget>[
+                getAppBarUI(),
+                main(),
+              ]),
         ),
       ),
     );
@@ -81,89 +73,92 @@ class _RegisterScreenState extends State<RegisterScreen> {
 
   Widget main() {
     return Expanded(
-    child: Padding(
-      padding: EdgeInsets.all(20),
-      child: SingleChildScrollView(
-        child: SizedBox(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.start,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: <Widget>[
-              Center(
-                child: Material(
-                  color: Colors.transparent,
-                  child: InkWell(
-                    borderRadius: const BorderRadius.all(Radius.circular(60.0)),
-                    child: Container(
-                      height: 120,
-                      width: 120,
-                      decoration: BoxDecoration(
-                        shape: BoxShape.circle,
-                        boxShadow: <BoxShadow>[
-                          BoxShadow(
-                            color: Theme.of(context).shadowColor, 
-                            offset: const Offset(2.0, 4.0), 
-                            blurRadius: 8),
-                        ],
+      child: Padding(
+        padding: EdgeInsets.all(20),
+        child: SingleChildScrollView(
+          child: SizedBox(
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.start,
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: <Widget>[
+                Center(
+                  child: Material(
+                    color: Colors.transparent,
+                    child: InkWell(
+                      borderRadius:
+                          const BorderRadius.all(Radius.circular(60.0)),
+                      child: Container(
+                        height: 120,
+                        width: 120,
+                        decoration: BoxDecoration(
+                          shape: BoxShape.circle,
+                          boxShadow: <BoxShadow>[
+                            BoxShadow(
+                                color: Theme.of(context).shadowColor,
+                                offset: const Offset(2.0, 4.0),
+                                blurRadius: 8),
+                          ],
+                        ),
+                        child: ClipRRect(
+                          borderRadius:
+                              const BorderRadius.all(Radius.circular(60.0)),
+                          child: photoFile != null
+                              ? Image.file(File(photoFile.path),
+                                  width: 120, height: 120, fit: BoxFit.fill)
+                              : Image.asset(
+                                  Helpers.getThemedThumbPlaceHolder(context)),
+                        ),
                       ),
-                      child: ClipRRect(
-                        borderRadius: const BorderRadius.all(Radius.circular(60.0)),
-                        child: photoFile != null ? Image.file(
-                          File(photoFile.path), 
-                          width: 120, 
-                          height: 120, 
-                          fit: BoxFit.fill
-                        ) : Image.asset(Helpers.getThemedThumbPlaceHolder(context)),
-                      ),
-                    ),
-                    onTap: () async {
-                      var pickedFile = await picker.getImage(source: ImageSource.gallery);
-                      print(pickedFile);
-                      setState(() {
-                        if (pickedFile != null) {
-                          photoFile = pickedFile;
-                        }
-                      });
-                    },
-                  ),
-                ),
-              ),
-
-              createTextTitle(tr("username")),
-              _buildUsername(),
-              createTextTitle(tr("password")),
-              _buildPassword(),
-              createTextTitle(tr("email")),
-              _buildEmail(),
-              createTextTitle(tr("full_name")),
-              _buildFullName(),
-              Padding(
-                padding: const EdgeInsets.only(top: 16),
-                child: Center(
-                  child: RoundedLoadingButton(
-                    width: MediaQuery.of(context).size.width,
-                    color: Theme.of(context).accentColor,
-                    controller: _btnController,
-                    onPressed: register,
-                    child: Text(
-                      tr("register"),
-                      style: TextStyle(
-                        fontWeight: FontWeight.w500,
-                        color: Theme.of(context).floatingActionButtonTheme.foregroundColor,
-                      ),
+                      onTap: () async {
+                        var pickedFile =
+                            await picker.getImage(source: ImageSource.gallery);
+                        print(pickedFile);
+                        setState(() {
+                          if (pickedFile != null) {
+                            photoFile = pickedFile;
+                          }
+                        });
+                      },
                     ),
                   ),
                 ),
-              ),
-              SizedBox(
-                height: 50,
-              ),
-            ],
+                createTextTitle(tr("username")),
+                _buildUsername(),
+                createTextTitle(tr("password")),
+                _buildPassword(),
+                createTextTitle(tr("email")),
+                _buildEmail(),
+                createTextTitle(tr("full_name")),
+                _buildFullName(),
+                Padding(
+                  padding: const EdgeInsets.only(top: 16),
+                  child: Center(
+                    child: RoundedLoadingButton(
+                      width: MediaQuery.of(context).size.width,
+                      color: Theme.of(context).accentColor,
+                      controller: _btnController,
+                      onPressed: register,
+                      child: Text(
+                        tr("register"),
+                        style: TextStyle(
+                          fontWeight: FontWeight.w500,
+                          color: Theme.of(context)
+                              .floatingActionButtonTheme
+                              .foregroundColor,
+                        ),
+                      ),
+                    ),
+                  ),
+                ),
+                SizedBox(
+                  height: 50,
+                ),
+              ],
+            ),
           ),
         ),
       ),
-    ),
-  );
+    );
   }
 
   Widget createTextTitle(String str) {
@@ -173,10 +168,9 @@ class _RegisterScreenState extends State<RegisterScreen> {
         str,
         textAlign: TextAlign.left,
         style: TextStyle(
-          fontSize: 20,
-          fontWeight: FontWeight.bold,
-          color: Theme.of(context).textTheme.subtitle1.color
-        ),
+            fontSize: 20,
+            fontWeight: FontWeight.bold,
+            color: Theme.of(context).textTheme.subtitle1?.color),
       ),
     );
   }
@@ -186,7 +180,7 @@ class _RegisterScreenState extends State<RegisterScreen> {
       padding: const EdgeInsets.only(top: 10, left: 0, right: 0),
       child: Container(
         decoration: BoxDecoration(
-          color: Theme.of(context).textTheme.bodyText1.color,
+          color: Theme.of(context).textTheme.bodyText1?.color,
           borderRadius: BorderRadius.circular(8),
         ),
         child: ClipRRect(
@@ -203,12 +197,11 @@ class _RegisterScreenState extends State<RegisterScreen> {
                 onChanged: (String txt) {},
                 style: TextStyle(
                   fontSize: 16,
-                  color: Theme.of(context).textTheme.caption.color,
+                  color: Theme.of(context).textTheme.caption?.color,
                 ),
                 cursorColor: Theme.of(context).accentColor,
                 decoration: InputDecoration(
-                    border: InputBorder.none,
-                    hintText: tr("enter_username")),
+                    border: InputBorder.none, hintText: tr("enter_username")),
               ),
             ),
           ),
@@ -222,7 +215,7 @@ class _RegisterScreenState extends State<RegisterScreen> {
       padding: const EdgeInsets.only(top: 10, left: 0, right: 0),
       child: Container(
         decoration: BoxDecoration(
-          color: Theme.of(context).textTheme.bodyText1.color,
+          color: Theme.of(context).textTheme.bodyText1?.color,
           borderRadius: BorderRadius.circular(8),
         ),
         child: ClipRRect(
@@ -239,12 +232,11 @@ class _RegisterScreenState extends State<RegisterScreen> {
                 onChanged: (String txt) {},
                 style: TextStyle(
                   fontSize: 16,
-                  color: Theme.of(context).textTheme.caption.color,
+                  color: Theme.of(context).textTheme.caption?.color,
                 ),
                 cursorColor: Theme.of(context).accentColor,
                 decoration: InputDecoration(
-                    border: InputBorder.none,
-                    hintText: tr("your_password")),
+                    border: InputBorder.none, hintText: tr("your_password")),
               ),
             ),
           ),
@@ -258,7 +250,7 @@ class _RegisterScreenState extends State<RegisterScreen> {
       padding: const EdgeInsets.only(top: 10, left: 0, right: 0),
       child: Container(
         decoration: BoxDecoration(
-          color: Theme.of(context).textTheme.bodyText1.color,
+          color: Theme.of(context).textTheme.bodyText1?.color,
           borderRadius: BorderRadius.circular(8),
         ),
         child: ClipRRect(
@@ -275,7 +267,7 @@ class _RegisterScreenState extends State<RegisterScreen> {
                 onChanged: (String txt) {},
                 style: TextStyle(
                   fontSize: 16,
-                  color: Theme.of(context).textTheme.caption.color,
+                  color: Theme.of(context).textTheme.caption?.color,
                 ),
                 cursorColor: Theme.of(context).accentColor,
                 decoration: InputDecoration(
@@ -294,7 +286,7 @@ class _RegisterScreenState extends State<RegisterScreen> {
       padding: const EdgeInsets.only(top: 10, left: 0, right: 0),
       child: Container(
         decoration: BoxDecoration(
-          color: Theme.of(context).textTheme.bodyText1.color,
+          color: Theme.of(context).textTheme.bodyText1?.color,
           borderRadius: BorderRadius.circular(8),
         ),
         child: ClipRRect(
@@ -311,7 +303,7 @@ class _RegisterScreenState extends State<RegisterScreen> {
                 onChanged: (String txt) {},
                 style: TextStyle(
                   fontSize: 16,
-                  color: Theme.of(context).textTheme.caption.color,
+                  color: Theme.of(context).textTheme.caption?.color,
                 ),
                 cursorColor: Theme.of(context).accentColor,
                 decoration: InputDecoration(
@@ -355,7 +347,6 @@ class _RegisterScreenState extends State<RegisterScreen> {
                 ),
               ),
             ),
-            
             Container(
               width: AppBar().preferredSize.height + 40,
               height: AppBar().preferredSize.height,
