@@ -1,23 +1,20 @@
-
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:directorio_delicias/main.dart';
 import 'package:directorio_delicias/themes/app_theme.dart';
 
-
 class SideMenuScreen extends StatefulWidget {
+  SideMenuScreen({Key? key, this.onCallbackFullScreen, this.sideMenus})
+      : super(key: key);
 
-  SideMenuScreen({Key key, this.onCallbackFullScreen, this.sideMenus}) : super(key: key);
-
-  final Function(DrawerEntry) onCallbackFullScreen;
-  final List<DrawerEntry> sideMenus;
+  final Function(DrawerEntry)? onCallbackFullScreen;
+  final List<DrawerEntry>? sideMenus;
 
   @override
   _SideMenuScreenState createState() => _SideMenuScreenState();
 }
 
 class _SideMenuScreenState extends State<SideMenuScreen> {
-  
   @override
   void initState() {
     super.initState();
@@ -33,46 +30,47 @@ class _SideMenuScreenState extends State<SideMenuScreen> {
           child: ListView.builder(
             padding: EdgeInsets.all(0),
             scrollDirection: Axis.vertical,
-            itemCount: MyApp.loggedUser != null ? widget.sideMenus.length : widget.sideMenus.length - 1,
+            itemCount: MyApp.loggedUser != null
+                ? widget.sideMenus?.length
+                : widget.sideMenus!.length - 1,
             itemBuilder: (context, index) {
               return Material(
                 color: Colors.transparent,
                 child: InkWell(
                   splashColor: Theme.of(context).accentColor,
                   child: Container(
-                    padding: EdgeInsets.only(left: 20, top: 10, right: 10, bottom: 10),
+                    padding: EdgeInsets.only(
+                        left: 20, top: 10, right: 10, bottom: 10),
                     height: 65,
                     alignment: Alignment.centerLeft,
                     child: Row(
                       children: [
                         Icon(
-                          widget.sideMenus[index].iconData,
+                          widget.sideMenus![index].iconData,
                           color: Colors.white,
                         ),
                         SizedBox(width: 10),
                         Text(
-                          widget.sideMenus[index].labelName,
+                          widget.sideMenus![index].labelName,
                           textAlign: TextAlign.start,
                           style: TextStyle(
-                            fontSize: 20,
-                            fontWeight: FontWeight.w400,
-                            color: Colors.white
-                          ),
+                              fontSize: 20,
+                              fontWeight: FontWeight.w400,
+                              color: Colors.white),
                         )
                       ],
                     ),
                   ),
                   onTap: () {
-                    DrawerEntry drawerEntry = widget.sideMenus[index];
-                    if(drawerEntry.isFullScreen) {
-                      widget.onCallbackFullScreen(drawerEntry);
+                    DrawerEntry drawerEntry = widget.sideMenus![index];
+                    if (drawerEntry.isFullScreen) {
+                      widget.onCallbackFullScreen!(drawerEntry);
                       print(drawerEntry.index);
-                    }
-                    else {
-                      MenuProvider provider = Provider.of<MenuProvider>(context, listen: false);
-                      if(provider.currentPage == index) 
-                        return;
-                        
+                    } else {
+                      MenuProvider provider =
+                          Provider.of<MenuProvider>(context, listen: false);
+                      if (provider.currentPage == index) return;
+
                       provider.updateCurrentPage(index);
                     }
                   },
@@ -84,7 +82,6 @@ class _SideMenuScreenState extends State<SideMenuScreen> {
       ),
     );
   }
-
 }
 
 enum MenuIndex {
@@ -105,20 +102,18 @@ enum MenuIndex {
 }
 
 class DrawerEntry {
-  DrawerEntry({
-    this.isAssetsImage = false,
-    this.labelName = '',
-    this.iconData,
-    this.index,
-    this.imageName = '',
-    this.isFullScreen = false
-  });
+  DrawerEntry(
+      {this.isAssetsImage = false,
+      this.labelName = '',
+      this.iconData,
+      this.index,
+      this.imageName = '',
+      this.isFullScreen = false});
 
   String labelName;
-  IconData iconData;
+  IconData? iconData;
   bool isAssetsImage;
   String imageName;
-  MenuIndex index;
+  MenuIndex? index;
   bool isFullScreen;
 }
-
