@@ -1,4 +1,3 @@
-
 import 'dart:async';
 import 'dart:collection';
 import 'dart:io';
@@ -27,8 +26,7 @@ import 'package:directorio_delicias/ui/widgets/load_widget.dart';
 import 'package:direct_select/direct_select.dart';
 
 class StoreAddScreen extends StatefulWidget {
-
-  StoreAddScreen({Key key}) : super(key: key);
+  StoreAddScreen({Key? key}) : super(key: key);
 
   @override
   _StoreAddScreenState createState() => _StoreAddScreenState();
@@ -37,11 +35,11 @@ class StoreAddScreen extends StatefulWidget {
 class _StoreAddScreenState extends State<StoreAddScreen> {
   @override
   void initState() {
-    
     super.initState();
   }
 
-  final RoundedLoadingButtonController _btnController = new RoundedLoadingButtonController();
+  final RoundedLoadingButtonController _btnController =
+      new RoundedLoadingButtonController();
   final TextEditingController _ctrlStoreName = new TextEditingController();
   final TextEditingController _ctrlStoreAddress = new TextEditingController();
   final TextEditingController _ctrlStoreDesc = new TextEditingController();
@@ -51,98 +49,92 @@ class _StoreAddScreenState extends State<StoreAddScreen> {
   final TextEditingController _ctrlWebsite = new TextEditingController();
 
   TimeOfDay _timeStartMon = TimeOfDay(hour: 0, minute: 0);
-  TimeOfDay _timeEndMon = TimeOfDay(hour: 23,minute: 59);
+  TimeOfDay _timeEndMon = TimeOfDay(hour: 23, minute: 59);
   TimeOfDay _timeStartTue = TimeOfDay(hour: 0, minute: 0);
-  TimeOfDay _timeEndTue = TimeOfDay(hour: 23,minute: 59);
+  TimeOfDay _timeEndTue = TimeOfDay(hour: 23, minute: 59);
   TimeOfDay _timeStartWed = TimeOfDay(hour: 0, minute: 0);
-  TimeOfDay _timeEndWed = TimeOfDay(hour: 23,minute: 59);
+  TimeOfDay _timeEndWed = TimeOfDay(hour: 23, minute: 59);
   TimeOfDay _timeStartThu = TimeOfDay(hour: 0, minute: 0);
-  TimeOfDay _timeEndThu = TimeOfDay(hour: 23,minute: 59);
+  TimeOfDay _timeEndThu = TimeOfDay(hour: 23, minute: 59);
   TimeOfDay _timeStartFri = TimeOfDay(hour: 0, minute: 0);
-  TimeOfDay _timeEndFri = TimeOfDay(hour: 23,minute: 59);
+  TimeOfDay _timeEndFri = TimeOfDay(hour: 23, minute: 59);
   TimeOfDay _timeStartSat = TimeOfDay(hour: 0, minute: 0);
-  TimeOfDay _timeEndSat = TimeOfDay(hour: 23,minute: 59);
+  TimeOfDay _timeEndSat = TimeOfDay(hour: 23, minute: 59);
   TimeOfDay _timeStartSun = TimeOfDay(hour: 0, minute: 0);
-  TimeOfDay _timeEndSun = TimeOfDay(hour: 23,minute: 59);
+  TimeOfDay _timeEndSun = TimeOfDay(hour: 23, minute: 59);
   Completer<GoogleMapController> _controller = Completer();
 
   var storageService = locator<SessionStorage>();
   // final _formKey = GlobalKey<FormState>();
-  List<Category> categories;
+  List<Category>? categories;
   int selectedCategoryIndex = 0;
-  Widget root;
-  Uint8List imageBytes;
-  LatLng selectedLatLng;
-  List<String> photoPaths = List<String>();
+  Widget? root;
+  Uint8List? imageBytes;
+  LatLng? selectedLatLng;
+  List<String> photoPaths = <String>[];
   final picker = ImagePicker();
-  User loggedUser;
+  User? loggedUser;
 
   void send() async {
     FocusScope.of(context).unfocus();
-    if(_ctrlStoreName.text.length == 0) {
+    if (_ctrlStoreName.text.length == 0) {
       Helpers.showAlertDialog(
-        context: context,
-        message: tr("store_name_empty"),
-        title: tr("action_error"),
-        onTapOk: null
-      );
+          context: context,
+          message: tr("store_name_empty"),
+          title: tr("action_error"),
+          onTapOk: null);
       return;
-    }
-    else if(selectedLatLng == null) {
+    } else if (selectedLatLng == null) {
       Helpers.showAlertDialog(
-        context: context,
-        title: tr("location_error"),
-        message: tr("location_error_details"),
-        onTapOk: null
-      );
+          context: context,
+          title: tr("location_error"),
+          message: tr("location_error_details"),
+          onTapOk: null);
       return;
-    }
-    else if(photoPaths.length == 0) {
+    } else if (photoPaths.length == 0) {
       Helpers.showAlertDialog(
-        context: context,
-        title: tr("photo_error"),
-        message: tr("photo_error_details"),
-        onTapOk: null
-      );
+          context: context,
+          title: tr("photo_error"),
+          message: tr("photo_error_details"),
+          onTapOk: null);
       return;
     }
 
     StoreServices.addStore(
-      storeId: "0", 
-      storeName: _ctrlStoreName.text, 
-      storeAddress: _ctrlStoreAddress.text, 
-      storeDesc: _ctrlStoreDesc.text, 
-      lat: selectedLatLng.latitude.toString(), 
-      lon: selectedLatLng.longitude.toString(), 
-      smsNo: _ctrlSMSNo.text, 
-      phoneNo: _ctrlPhoneNo.text, 
-      email: _ctrlEmail.text, 
-      website: _ctrlWebsite.text, 
-      categoryId: categories[selectedCategoryIndex].categoryId.toString(), 
-      monOpen: _timeStartMon, 
-      monClose: _timeEndMon, 
-      tueOpen: _timeStartTue, 
-      tueClose: _timeEndTue, 
-      wedOpen: _timeStartWed, 
-      wedClose: _timeEndWed,  
-      thuOpen: _timeStartThu, 
-      thuClose: _timeEndThu,
-      friOpen: _timeStartFri, 
-      friClose: _timeEndFri, 
-      satOpen: _timeStartSat, 
-      satClose: _timeEndSat, 
-      sunOpen: _timeStartSun, 
-      sunClose: _timeEndSun,
-      userId: loggedUser.userId, 
-      loginHash: loggedUser.loginHash,
-      photoPaths: photoPaths,
-      context: context,
-      photoIdsDeleted: List<int>()
-    );
+        storeId: "0",
+        storeName: _ctrlStoreName.text,
+        storeAddress: _ctrlStoreAddress.text,
+        storeDesc: _ctrlStoreDesc.text,
+        lat: selectedLatLng!.latitude.toString(),
+        lon: selectedLatLng!.longitude.toString(),
+        smsNo: _ctrlSMSNo.text,
+        phoneNo: _ctrlPhoneNo.text,
+        email: _ctrlEmail.text,
+        website: _ctrlWebsite.text,
+        categoryId: categories![selectedCategoryIndex].categoryId.toString(),
+        monOpen: _timeStartMon,
+        monClose: _timeEndMon,
+        tueOpen: _timeStartTue,
+        tueClose: _timeEndTue,
+        wedOpen: _timeStartWed,
+        wedClose: _timeEndWed,
+        thuOpen: _timeStartThu,
+        thuClose: _timeEndThu,
+        friOpen: _timeStartFri,
+        friClose: _timeEndFri,
+        satOpen: _timeStartSat,
+        satClose: _timeEndSat,
+        sunOpen: _timeStartSun,
+        sunClose: _timeEndSun,
+        userId: loggedUser!.userId,
+        loginHash: loggedUser!.loginHash,
+        photoPaths: photoPaths,
+        context: context,
+        photoIdsDeleted: <int>[]);
   }
 
   List<Widget> _buildItems() {
-    return categories
+    return categories!
         .map((val) => CategorySelectionItem(
               category: val,
             ))
@@ -164,17 +156,16 @@ class _StoreAddScreenState extends State<StoreAddScreen> {
         child: Scaffold(
           backgroundColor: Theme.of(context).backgroundColor,
           body: Column(
-            mainAxisAlignment: MainAxisAlignment.start,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: <Widget>[
-              getAppBarUI(),
-              main(),
-            ]
-          ),
+              mainAxisAlignment: MainAxisAlignment.start,
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: <Widget>[
+                getAppBarUI(),
+                main(),
+              ]),
         ),
       ),
     );
-    return root;
+    return root!;
   }
 
   Widget googleMap() {
@@ -184,54 +175,53 @@ class _StoreAddScreenState extends State<StoreAddScreen> {
     );
 
     Marker marker = Marker(
-      consumeTapEvents: false,
-      markerId: MarkerId("CURRENT_LOCATION"),
-      position: LatLng(Config.DEFAULT_LAT, Config.DEFAULT_LON),
-      anchor: Offset(0, -0.75),
-      icon: BitmapDescriptor.defaultMarker,
-      infoWindow: InfoWindow(
-        title: tr("tap_and_hold"),
-        snippet: tr("drag_me_around"),
-      )
-    );
+        consumeTapEvents: false,
+        markerId: MarkerId("CURRENT_LOCATION"),
+        position: LatLng(Config.DEFAULT_LAT, Config.DEFAULT_LON),
+        anchor: Offset(0, -0.75),
+        icon: BitmapDescriptor.defaultMarker,
+        infoWindow: InfoWindow(
+          title: tr("tap_and_hold"),
+          snippet: tr("drag_me_around"),
+        ));
 
     Set<Marker> markers = HashSet<Marker>();
     markers.add(marker);
-    
+
     return ClipRRect(
-      borderRadius: BorderRadius.all(Radius.circular(10)),
-      child: GoogleMap(
-        gestureRecognizers: <foundation.Factory<OneSequenceGestureRecognizer>>[
-          new foundation.Factory<OneSequenceGestureRecognizer>(
-            () => new EagerGestureRecognizer(),
-          ),
-        ].toSet(),
-        zoomControlsEnabled: false,
-        zoomGesturesEnabled: false,
-        compassEnabled: false,
-        mapToolbarEnabled: false,
-        scrollGesturesEnabled: false,
-        rotateGesturesEnabled: false,
-        tiltGesturesEnabled: false,
-        markers: markers,
-        myLocationButtonEnabled: false,
-        mapType: MapType.normal,
-        initialCameraPosition: cameraPosition,
-        onMapCreated: (GoogleMapController controller) {
-          _controller.complete(controller);
-          controller.animateCamera(CameraUpdate.newCameraPosition(cameraPosition));
-          controller.showMarkerInfoWindow(marker.markerId);
-        },
-      )
-    );
+        borderRadius: BorderRadius.all(Radius.circular(10)),
+        child: GoogleMap(
+          gestureRecognizers:
+              <foundation.Factory<OneSequenceGestureRecognizer>>[
+            new foundation.Factory<OneSequenceGestureRecognizer>(
+              () => new EagerGestureRecognizer(),
+            ),
+          ].toSet(),
+          zoomControlsEnabled: false,
+          zoomGesturesEnabled: false,
+          compassEnabled: false,
+          mapToolbarEnabled: false,
+          scrollGesturesEnabled: false,
+          rotateGesturesEnabled: false,
+          tiltGesturesEnabled: false,
+          markers: markers,
+          myLocationButtonEnabled: false,
+          mapType: MapType.normal,
+          initialCameraPosition: cameraPosition,
+          onMapCreated: (GoogleMapController controller) {
+            _controller.complete(controller);
+            controller
+                .animateCamera(CameraUpdate.newCameraPosition(cameraPosition));
+            controller.showMarkerInfoWindow(marker.markerId);
+          },
+        ));
   }
 
-  
   @override
   Widget build(BuildContext context) {
     return FutureBuilder<bool>(
       future: getData(),
-      builder:(BuildContext context, AsyncSnapshot<bool> snapshot) {
+      builder: (BuildContext context, AsyncSnapshot<bool> snapshot) {
         if (!snapshot.hasData) {
           return LoadingWidget(
             size: 30.0,
@@ -246,44 +236,42 @@ class _StoreAddScreenState extends State<StoreAddScreen> {
 
   Widget main() {
     var isDark = Theme.of(context).brightness == Brightness.dark;
-    String mapPlaceholder = isDark ? Config.DEFAULT_MAP_DARK_PLACEHOLDER : Config.DEFAULT_MAP_PLACEHOLDER;
+    String mapPlaceholder = isDark
+        ? Config.DEFAULT_MAP_DARK_PLACEHOLDER
+        : Config.DEFAULT_MAP_PLACEHOLDER;
 
     var categoryWidget;
-    if(selectedCategoryIndex < categories.length) {
+    if (selectedCategoryIndex < categories!.length) {
       categoryWidget = DirectSelect(
-        itemExtent: 50.0,
-        mode: DirectSelectMode.tap,
-        selectedIndex: selectedCategoryIndex,
-        backgroundColor: Theme.of(context).accentColor,
-        child: CategorySelectionItem(
-          isForList: false,
-          category: categories[selectedCategoryIndex],
-        ),
-        onSelectedItemChanged: (index) {
-          setState(() {
-            selectedCategoryIndex = index;
-          });
-        },
-        items: _buildItems()
-      );
-    }
-    else {
+          itemExtent: 50.0,
+          mode: DirectSelectMode.tap,
+          selectedIndex: selectedCategoryIndex,
+          backgroundColor: Theme.of(context).accentColor,
+          child: CategorySelectionItem(
+            isForList: false,
+            category: categories![selectedCategoryIndex],
+          ),
+          onSelectedItemChanged: (index) {
+            setState(() {
+              selectedCategoryIndex = index!;
+            });
+          },
+          items: _buildItems());
+    } else {
       categoryWidget = Helpers.createFormField(
-        controller: _ctrlStoreAddress,
-        placeholder: tr("no_categories_to_show"),
-        validatorText: null,
-        context: context,
-        enabled: false
-      );
+          controller: _ctrlStoreAddress,
+          placeholder: tr("no_categories_to_show"),
+          validatorText: null,
+          context: context,
+          enabled: false);
     }
 
     double leftPhoto = 0.0;
     double rightPhoto = 0.0;
 
-    if(Helpers.isRTL()) {
+    if (Helpers.isRTL()) {
       rightPhoto = 8.0;
-    }
-    else {
+    } else {
       leftPhoto = 8.0;
     }
 
@@ -301,62 +289,55 @@ class _StoreAddScreenState extends State<StoreAddScreen> {
                   text: tr("store_name"),
                 ),
                 Helpers.createFormField(
-                  controller: _ctrlStoreName,
-                  placeholder: tr("store_name"),
-                  // validatorText: tr("store_name_empty"),
-                  context: context
-                ),
+                    controller: _ctrlStoreName,
+                    placeholder: tr("store_name"),
+                    // validatorText: tr("store_name_empty"),
+                    context: context),
                 Helpers.createHeader(
                   context: context,
                   text: tr("store_address"),
                 ),
                 Helpers.createFormField(
-                  controller: _ctrlStoreAddress,
-                  placeholder: tr("store_address"),
-                  validatorText: null,
-                  context: context
-                ),
+                    controller: _ctrlStoreAddress,
+                    placeholder: tr("store_address"),
+                    validatorText: null,
+                    context: context),
                 Helpers.createHeader(
                   context: context,
                   text: tr("sms_no"),
                 ),
                 Helpers.createFormField(
-                  controller: _ctrlSMSNo,
-                  placeholder: tr("sms_no"),
-                  validatorText: null,
-                  context: context
-                ),
+                    controller: _ctrlSMSNo,
+                    placeholder: tr("sms_no"),
+                    validatorText: null,
+                    context: context),
                 Helpers.createHeader(
                   context: context,
                   text: tr("phone_no"),
                 ),
                 Helpers.createFormField(
-                  controller: _ctrlPhoneNo,
-                  placeholder: tr("phone_no"),
-                  validatorText: null,
-                  context: context
-                ),
-
+                    controller: _ctrlPhoneNo,
+                    placeholder: tr("phone_no"),
+                    validatorText: null,
+                    context: context),
                 Helpers.createHeader(
                   context: context,
                   text: tr("email"),
                 ),
                 Helpers.createFormField(
-                  controller: _ctrlEmail,
-                  placeholder: tr("email"),
-                  validatorText: null,
-                  context: context
-                ),
+                    controller: _ctrlEmail,
+                    placeholder: tr("email"),
+                    validatorText: null,
+                    context: context),
                 Helpers.createHeader(
                   context: context,
                   text: tr("website"),
                 ),
                 Helpers.createFormField(
-                  controller: _ctrlWebsite,
-                  placeholder: tr("website"),
-                  validatorText: null,
-                  context: context
-                ),
+                    controller: _ctrlWebsite,
+                    placeholder: tr("website"),
+                    validatorText: null,
+                    context: context),
                 Helpers.createHeader(
                   context: context,
                   text: tr("category"),
@@ -367,196 +348,195 @@ class _StoreAddScreenState extends State<StoreAddScreen> {
                   text: tr("desc"),
                 ),
                 Helpers.createFormField(
-                  controller: _ctrlStoreDesc,
-                  placeholder: tr("desc"),
-                  validatorText: null,
-                  context: context,
-                  height: 150
+                    controller: _ctrlStoreDesc,
+                    placeholder: tr("desc"),
+                    validatorText: null,
+                    context: context,
+                    height: 150),
+                SizedBox(
+                  height: 50,
                 ),
-                SizedBox(height: 50,),
                 Helpers.createHeader(
                   context: context,
                   text: tr("operating_hours"),
                 ),
-                SizedBox(height: 20,),
+                SizedBox(
+                  height: 20,
+                ),
                 Helpers.createHeader(
                   context: context,
                   text: tr("mon"),
                 ),
                 Helpers.createTimeRange(
-                  context: context, 
-                  startTime: _timeStartMon, 
-                  endTime: _timeEndMon,
-                  onTimeStartChanged: (newTime) {
-                    setState(() {
-                      _timeStartMon = newTime;
-                    });
-                  },
-                  onTimeEndChanged: (newTime) {
-                    setState(() {
-                      _timeEndMon = newTime;
-                    });
-                  }
-                ),
+                    context: context,
+                    startTime: _timeStartMon,
+                    endTime: _timeEndMon,
+                    onTimeStartChanged: (newTime) {
+                      setState(() {
+                        _timeStartMon = newTime;
+                      });
+                    },
+                    onTimeEndChanged: (newTime) {
+                      setState(() {
+                        _timeEndMon = newTime;
+                      });
+                    }),
                 Helpers.createHeader(
                   context: context,
                   text: tr("tue"),
                 ),
                 Helpers.createTimeRange(
-                  context: context, 
-                  startTime: _timeStartTue, 
-                  endTime: _timeEndTue,
-                  onTimeStartChanged: (newTime) {
-                    setState(() {
-                      _timeStartTue = newTime;
-                    });
-                  },
-                  onTimeEndChanged: (newTime) {
-                    setState(() {
-                      _timeEndTue = newTime;
-                    });
-                  }
-                ),
+                    context: context,
+                    startTime: _timeStartTue,
+                    endTime: _timeEndTue,
+                    onTimeStartChanged: (newTime) {
+                      setState(() {
+                        _timeStartTue = newTime;
+                      });
+                    },
+                    onTimeEndChanged: (newTime) {
+                      setState(() {
+                        _timeEndTue = newTime;
+                      });
+                    }),
                 Helpers.createHeader(
                   context: context,
                   text: tr("wed"),
                 ),
                 Helpers.createTimeRange(
-                  context: context, 
-                  startTime: _timeStartWed, 
-                  endTime: _timeEndWed,
-                  onTimeStartChanged: (newTime) {
-                    setState(() {
-                      _timeStartWed = newTime;
-                    });
-                  },
-                  onTimeEndChanged: (newTime) {
-                    setState(() {
-                      _timeEndWed = newTime;
-                    });
-                  }
-                ),
+                    context: context,
+                    startTime: _timeStartWed,
+                    endTime: _timeEndWed,
+                    onTimeStartChanged: (newTime) {
+                      setState(() {
+                        _timeStartWed = newTime;
+                      });
+                    },
+                    onTimeEndChanged: (newTime) {
+                      setState(() {
+                        _timeEndWed = newTime;
+                      });
+                    }),
                 Helpers.createHeader(
                   context: context,
                   text: tr("thu"),
                 ),
                 Helpers.createTimeRange(
-                  context: context, 
-                  startTime: _timeStartThu, 
-                  endTime: _timeEndThu,
-                  onTimeStartChanged: (newTime) {
-                    setState(() {
-                      _timeStartThu = newTime;
-                    });
-                  },
-                  onTimeEndChanged: (newTime) {
-                    setState(() {
-                      _timeEndThu = newTime;
-                    });
-                  }
-                ),
+                    context: context,
+                    startTime: _timeStartThu,
+                    endTime: _timeEndThu,
+                    onTimeStartChanged: (newTime) {
+                      setState(() {
+                        _timeStartThu = newTime;
+                      });
+                    },
+                    onTimeEndChanged: (newTime) {
+                      setState(() {
+                        _timeEndThu = newTime;
+                      });
+                    }),
                 Helpers.createHeader(
                   context: context,
                   text: tr("fri"),
                 ),
                 Helpers.createTimeRange(
-                  context: context, 
-                  startTime: _timeStartFri, 
-                  endTime: _timeEndFri,
-                  onTimeStartChanged: (newTime) {
-                    setState(() {
-                      _timeStartFri = newTime;
-                    });
-                  },
-                  onTimeEndChanged: (newTime) {
-                    setState(() {
-                      _timeEndFri = newTime;
-                    });
-                  }
-                ),
+                    context: context,
+                    startTime: _timeStartFri,
+                    endTime: _timeEndFri,
+                    onTimeStartChanged: (newTime) {
+                      setState(() {
+                        _timeStartFri = newTime;
+                      });
+                    },
+                    onTimeEndChanged: (newTime) {
+                      setState(() {
+                        _timeEndFri = newTime;
+                      });
+                    }),
                 Helpers.createHeader(
                   context: context,
                   text: tr("sat"),
                 ),
                 Helpers.createTimeRange(
-                  context: context, 
-                  startTime: _timeStartSat, 
-                  endTime: _timeEndSat,
-                  onTimeStartChanged: (newTime) {
-                    setState(() {
-                      _timeStartSat = newTime;
-                    });
-                  },
-                  onTimeEndChanged: (newTime) {
-                    setState(() {
-                      _timeEndSat = newTime;
-                    });
-                  }
-                ),
+                    context: context,
+                    startTime: _timeStartSat,
+                    endTime: _timeEndSat,
+                    onTimeStartChanged: (newTime) {
+                      setState(() {
+                        _timeStartSat = newTime;
+                      });
+                    },
+                    onTimeEndChanged: (newTime) {
+                      setState(() {
+                        _timeEndSat = newTime;
+                      });
+                    }),
                 Helpers.createHeader(
                   context: context,
                   text: tr("sun"),
                 ),
                 Helpers.createTimeRange(
-                  context: context, 
-                  startTime: _timeStartSun, 
-                  endTime: _timeEndSun,
-                  onTimeStartChanged: (newTime) {
-                    setState(() {
-                      _timeStartSun = newTime;
-                    });
-                  },
-                  onTimeEndChanged: (newTime) {
-                    setState(() {
-                      _timeEndSun = newTime;
-                    });
-                  }
+                    context: context,
+                    startTime: _timeStartSun,
+                    endTime: _timeEndSun,
+                    onTimeStartChanged: (newTime) {
+                      setState(() {
+                        _timeStartSun = newTime;
+                      });
+                    },
+                    onTimeEndChanged: (newTime) {
+                      setState(() {
+                        _timeEndSun = newTime;
+                      });
+                    }),
+                SizedBox(
+                  height: 50,
                 ),
-                SizedBox(height: 50,),
                 Helpers.createHeader(
                   context: context,
                   text: tr("store_location"),
                 ),
-                SizedBox(height: 10,),
-                InkWell(
-                  onTap: () async{
-                    FocusScope.of(context).unfocus();
-                    DataHandler dataHandler = await Navigator.push<dynamic>(
-                      context,
-                      MaterialPageRoute<dynamic>(
-                          builder: (BuildContext context) => StoreMapScreen(),
-                          fullscreenDialog: true),
-                    );
-                    if(dataHandler != null) {
-                      setState(() {
-                        imageBytes = dataHandler.bytes;
-                        selectedLatLng = dataHandler.latLng;
-                      });
-                    }
-                  },
-                  child: Container(
-                    height: 200,
-                    width: MediaQuery.of(context).size.width,
-                    child: ClipRRect(
-                      borderRadius: BorderRadius.circular(8),
-                      child: AspectRatio(
-                        aspectRatio: 1.2,
-                        child: imageBytes != null ?
-                          FadeInImage(
-                            fit: BoxFit.cover,
-                            placeholder: AssetImage(mapPlaceholder),
-                            image: MemoryImage(imageBytes),
-                          ) :
-                          FadeInImage(
-                            fit: BoxFit.cover,
-                            placeholder: AssetImage(mapPlaceholder),
-                            image: AssetImage(mapPlaceholder),
-                          )
-                      ),
-                    )
-                  )
+                SizedBox(
+                  height: 10,
                 ),
-                SizedBox(height: 50,),
+                InkWell(
+                    onTap: () async {
+                      FocusScope.of(context).unfocus();
+                      DataHandler dataHandler = await Navigator.push<dynamic>(
+                        context,
+                        MaterialPageRoute<dynamic>(
+                            builder: (BuildContext context) => StoreMapScreen(),
+                            fullscreenDialog: true),
+                      );
+                      if (dataHandler != null) {
+                        setState(() {
+                          imageBytes = dataHandler.bytes;
+                          selectedLatLng = dataHandler.latLng;
+                        });
+                      }
+                    },
+                    child: Container(
+                        height: 200,
+                        width: MediaQuery.of(context).size.width,
+                        child: ClipRRect(
+                          borderRadius: BorderRadius.circular(8),
+                          child: AspectRatio(
+                              aspectRatio: 1.2,
+                              child: imageBytes != null
+                                  ? FadeInImage(
+                                      fit: BoxFit.cover,
+                                      placeholder: AssetImage(mapPlaceholder),
+                                      image: MemoryImage(imageBytes!),
+                                    )
+                                  : FadeInImage(
+                                      fit: BoxFit.cover,
+                                      placeholder: AssetImage(mapPlaceholder),
+                                      image: AssetImage(mapPlaceholder),
+                                    )),
+                        ))),
+                SizedBox(
+                  height: 50,
+                ),
                 Helpers.createHeader(
                   context: context,
                   text: tr("store_photos"),
@@ -570,7 +550,8 @@ class _StoreAddScreenState extends State<StoreAddScreen> {
                         Expanded(
                           child: Container(
                             decoration: BoxDecoration(
-                              color: Theme.of(context).textTheme.bodyText1.color,
+                              color:
+                                  Theme.of(context).textTheme.bodyText1?.color,
                               borderRadius: BorderRadius.circular(8),
                             ),
                             child: ListView.builder(
@@ -579,35 +560,39 @@ class _StoreAddScreenState extends State<StoreAddScreen> {
                               itemBuilder: (BuildContext context, int index) {
                                 return InkWell(
                                   child: Padding(
-                                    padding: EdgeInsets.only(left: leftPhoto, top: 8.0, bottom: 8.0, right: rightPhoto ),
-                                    child: Container(
-                                      height: 140,
-                                      width: 140,
-                                      child: ClipRRect(
-                                        borderRadius: BorderRadius.circular(8),
-                                        child: AspectRatio(
-                                          aspectRatio: 1.2,
-                                          child: FadeInImage(
-                                            fit: BoxFit.cover,
-                                            placeholder: AssetImage(Helpers.getThemedPhotoPlaceHolder(context)),
-                                            image: FileImage(File(photoPaths[index])),
-                                          )
-                                        ),
-                                      )
-                                    )
-                                  ),
+                                      padding: EdgeInsets.only(
+                                          left: leftPhoto,
+                                          top: 8.0,
+                                          bottom: 8.0,
+                                          right: rightPhoto),
+                                      child: Container(
+                                          height: 140,
+                                          width: 140,
+                                          child: ClipRRect(
+                                            borderRadius:
+                                                BorderRadius.circular(8),
+                                            child: AspectRatio(
+                                                aspectRatio: 1.2,
+                                                child: FadeInImage(
+                                                  fit: BoxFit.cover,
+                                                  placeholder: AssetImage(Helpers
+                                                      .getThemedPhotoPlaceHolder(
+                                                          context)),
+                                                  image: FileImage(
+                                                      File(photoPaths[index])),
+                                                )),
+                                          ))),
                                   onTap: () {
                                     Helpers.showDualButtonDialog(
-                                      context: context,
-                                      title: tr("delete_photo"),
-                                      message: tr("delete_photo_details"),
-                                      onTapDismiss: () { },
-                                      onTapOk: () {
-                                        setState(() {
-                                          photoPaths.removeAt(index);
+                                        context: context,
+                                        title: tr("delete_photo"),
+                                        message: tr("delete_photo_details"),
+                                        onTapDismiss: () {},
+                                        onTapOk: () {
+                                          setState(() {
+                                            photoPaths.removeAt(index);
+                                          });
                                         });
-                                      }
-                                    );
                                   },
                                 );
                               },
@@ -617,40 +602,45 @@ class _StoreAddScreenState extends State<StoreAddScreen> {
                         Padding(
                           padding: const EdgeInsets.only(left: 10, right: 10),
                           child: Container(
-                            width: 50,
-                            height: MediaQuery.of(context).size.height,
-                            decoration: BoxDecoration(
-                              color: Theme.of(context).accentColor,
-                              borderRadius: BorderRadius.circular(8),
-                            ),
-                            child: Material(
-                              color: Theme.of(context).accentColor,
-                              borderRadius: BorderRadius.circular(8),
-                              child: InkWell(
+                              width: 50,
+                              height: MediaQuery.of(context).size.height,
+                              decoration: BoxDecoration(
+                                color: Theme.of(context).accentColor,
                                 borderRadius: BorderRadius.circular(8),
-                                child: Icon(
-                                  Icons.add_a_photo,
-                                  color: Theme.of(context).floatingActionButtonTheme.foregroundColor,
-                                ),
-                                onTap: () async {
-                                  if(photoPaths.length == Config.MAX_PHOTO_UPLOAD) {
-                                    Helpers.showAlertDialog(
-                                      context: context,
-                                      title: tr("max_photos_reached"),
-                                      message: tr("max_photos_reached_details"),
-                                    );
-                                    return;
-                                  }
-                                  PickedFile pickedFile = await picker.getImage(source: ImageSource.gallery, imageQuality: 90);
-                                  setState(() {
-                                    if (pickedFile != null) {
-                                      photoPaths.add(pickedFile.path);
-                                    }
-                                  });
-                                }
                               ),
-                            )
-                          ),
+                              child: Material(
+                                color: Theme.of(context).accentColor,
+                                borderRadius: BorderRadius.circular(8),
+                                child: InkWell(
+                                    borderRadius: BorderRadius.circular(8),
+                                    child: Icon(
+                                      Icons.add_a_photo,
+                                      color: Theme.of(context)
+                                          .floatingActionButtonTheme
+                                          .foregroundColor,
+                                    ),
+                                    onTap: () async {
+                                      if (photoPaths.length ==
+                                          Config.MAX_PHOTO_UPLOAD) {
+                                        Helpers.showAlertDialog(
+                                          context: context,
+                                          title: tr("max_photos_reached"),
+                                          message:
+                                              tr("max_photos_reached_details"),
+                                        );
+                                        return;
+                                      }
+                                      PickedFile? pickedFile =
+                                          await picker.getImage(
+                                              source: ImageSource.gallery,
+                                              imageQuality: 90);
+                                      setState(() {
+                                        if (pickedFile != null) {
+                                          photoPaths.add(pickedFile.path);
+                                        }
+                                      });
+                                    }),
+                              )),
                         )
                       ],
                     ),
@@ -669,7 +659,9 @@ class _StoreAddScreenState extends State<StoreAddScreen> {
                         tr("add"),
                         style: TextStyle(
                           fontWeight: FontWeight.w500,
-                          color: Theme.of(context).floatingActionButtonTheme.foregroundColor,
+                          color: Theme.of(context)
+                              .floatingActionButtonTheme
+                              .foregroundColor,
                         ),
                       ),
                     ),
@@ -686,7 +678,6 @@ class _StoreAddScreenState extends State<StoreAddScreen> {
     );
   }
 
-  
   Widget getAppBarUI() {
     return Container(
       decoration: BoxDecoration(
@@ -738,6 +729,4 @@ class _StoreAddScreenState extends State<StoreAddScreen> {
       ),
     );
   }
-  
 }
-
