@@ -1,4 +1,3 @@
-
 import 'dart:async';
 import 'dart:io';
 import 'dart:typed_data';
@@ -26,48 +25,48 @@ import 'package:directorio_delicias/ui/widgets/load_widget.dart';
 import 'package:direct_select/direct_select.dart';
 
 class StoreEditScreen extends StatefulWidget {
+  StoreEditScreen({Key? key, this.store}) : super(key: key);
 
-  StoreEditScreen({Key key, this.store}) : super(key: key);
-  
-  final Store store;
+  final Store? store;
 
   @override
   _StoreEditScreenState createState() => _StoreEditScreenState();
 }
 
 class _StoreEditScreenState extends State<StoreEditScreen> {
-
   @override
   void initState() {
     super.initState();
-    for(Photo photo in widget.store.photos) {
+    for (Photo photo in widget.store!.photos!) {
       photoObjs.add(photo);
     }
 
-    _timeStartMon = Helpers.formatTimeOfDay(widget.store.monOpen);
-    _timeEndMon = Helpers.formatTimeOfDay(widget.store.monClose);
+    _timeStartMon = Helpers.formatTimeOfDay(widget.store!.monOpen);
+    _timeEndMon = Helpers.formatTimeOfDay(widget.store!.monClose);
 
-    _timeStartTue = Helpers.formatTimeOfDay(widget.store.tueOpen);
-    _timeEndTue = Helpers.formatTimeOfDay(widget.store.tueClose);
+    _timeStartTue = Helpers.formatTimeOfDay(widget.store!.tueOpen);
+    _timeEndTue = Helpers.formatTimeOfDay(widget.store!.tueClose);
 
-    _timeStartWed = Helpers.formatTimeOfDay(widget.store.wedOpen);
-    _timeEndWed = Helpers.formatTimeOfDay(widget.store.wedClose);
+    _timeStartWed = Helpers.formatTimeOfDay(widget.store!.wedOpen);
+    _timeEndWed = Helpers.formatTimeOfDay(widget.store!.wedClose);
 
-    _timeStartThu = Helpers.formatTimeOfDay(widget.store.thuOpen);
-    _timeEndThu = Helpers.formatTimeOfDay(widget.store.thuClose);
+    _timeStartThu = Helpers.formatTimeOfDay(widget.store!.thuOpen);
+    _timeEndThu = Helpers.formatTimeOfDay(widget.store!.thuClose);
 
-    _timeStartFri = Helpers.formatTimeOfDay(widget.store.friOpen);
-    _timeEndFri = Helpers.formatTimeOfDay(widget.store.friClose);
+    _timeStartFri = Helpers.formatTimeOfDay(widget.store!.friOpen);
+    _timeEndFri = Helpers.formatTimeOfDay(widget.store!.friClose);
 
-    _timeStartSat = Helpers.formatTimeOfDay(widget.store.satOpen);
-    _timeEndSat = Helpers.formatTimeOfDay(widget.store.satClose);
+    _timeStartSat = Helpers.formatTimeOfDay(widget.store!.satOpen);
+    _timeEndSat = Helpers.formatTimeOfDay(widget.store!.satClose);
 
-    _timeStartSun = Helpers.formatTimeOfDay(widget.store.sunOpen);
-    _timeEndSun = Helpers.formatTimeOfDay(widget.store.sunClose);
+    _timeStartSun = Helpers.formatTimeOfDay(widget.store!.sunOpen);
+    _timeEndSun = Helpers.formatTimeOfDay(widget.store!.sunClose);
   }
 
-  final RoundedLoadingButtonController _btnController = new RoundedLoadingButtonController();
-  final RoundedLoadingButtonController _btnDeleteController = new RoundedLoadingButtonController();
+  final RoundedLoadingButtonController _btnController =
+      new RoundedLoadingButtonController();
+  final RoundedLoadingButtonController _btnDeleteController =
+      new RoundedLoadingButtonController();
   final TextEditingController _ctrlStoreName = new TextEditingController();
   final TextEditingController _ctrlStoreAddress = new TextEditingController();
   final TextEditingController _ctrlStoreDesc = new TextEditingController();
@@ -77,123 +76,116 @@ class _StoreEditScreenState extends State<StoreEditScreen> {
   final TextEditingController _ctrlWebsite = new TextEditingController();
 
   TimeOfDay _timeStartMon = TimeOfDay(hour: 0, minute: 0);
-  TimeOfDay _timeEndMon = TimeOfDay(hour: 23,minute: 59);
+  TimeOfDay _timeEndMon = TimeOfDay(hour: 23, minute: 59);
   TimeOfDay _timeStartTue = TimeOfDay(hour: 0, minute: 0);
-  TimeOfDay _timeEndTue = TimeOfDay(hour: 23,minute: 59);
+  TimeOfDay _timeEndTue = TimeOfDay(hour: 23, minute: 59);
   TimeOfDay _timeStartWed = TimeOfDay(hour: 0, minute: 0);
-  TimeOfDay _timeEndWed = TimeOfDay(hour: 23,minute: 59);
+  TimeOfDay _timeEndWed = TimeOfDay(hour: 23, minute: 59);
   TimeOfDay _timeStartThu = TimeOfDay(hour: 0, minute: 0);
-  TimeOfDay _timeEndThu = TimeOfDay(hour: 23,minute: 59);
+  TimeOfDay _timeEndThu = TimeOfDay(hour: 23, minute: 59);
   TimeOfDay _timeStartFri = TimeOfDay(hour: 0, minute: 0);
-  TimeOfDay _timeEndFri = TimeOfDay(hour: 23,minute: 59);
+  TimeOfDay _timeEndFri = TimeOfDay(hour: 23, minute: 59);
   TimeOfDay _timeStartSat = TimeOfDay(hour: 0, minute: 0);
-  TimeOfDay _timeEndSat = TimeOfDay(hour: 23,minute: 59);
+  TimeOfDay _timeEndSat = TimeOfDay(hour: 23, minute: 59);
   TimeOfDay _timeStartSun = TimeOfDay(hour: 0, minute: 0);
-  TimeOfDay _timeEndSun = TimeOfDay(hour: 23,minute: 59);
+  TimeOfDay _timeEndSun = TimeOfDay(hour: 23, minute: 59);
 
   var storageService = locator<SessionStorage>();
-  List<Category> categories;
+  List<Category>? categories;
   int selectedCategoryIndex = 0;
-  Widget root;
-  Uint8List imageBytes;
-  LatLng selectedLatLng;
-  List<dynamic> photoObjs = List<dynamic>();
-  List<int> photoIdsDeleted = List<int>();
+  Widget? root;
+  Uint8List? imageBytes;
+  LatLng? selectedLatLng;
+  List<dynamic> photoObjs = <dynamic>[];
+  List<int> photoIdsDeleted = <int>[];
   final picker = ImagePicker();
-  User loggedUser;
+  User? loggedUser;
 
   void deleteStore() async {
     FocusScope.of(context).unfocus();
     Helpers.showDualButtonDialog(
-        context: context,
-        message: tr("delete_store_details"),
-        title: tr("delete_store"),
-        onTapDismiss: () { },
-        onTapOk: () {
-            
-            Future.delayed(Duration(milliseconds: 500), () {
-                        proceedDelete();
-            });
-        },
-      );
-      print(photoObjs.length);
+      context: context,
+      message: tr("delete_store_details"),
+      title: tr("delete_store"),
+      onTapDismiss: () {},
+      onTapOk: () {
+        Future.delayed(Duration(milliseconds: 500), () {
+          proceedDelete();
+        });
+      },
+    );
+    print(photoObjs.length);
   }
 
   void proceedDelete() async {
     StoreServices.deleteStore(
-      storeId: widget.store.storeId, 
-      userId: loggedUser.userId, 
-      loginHash: loggedUser.loginHash,
+      storeId: widget.store!.storeId,
+      userId: loggedUser!.userId,
+      loginHash: loggedUser!.loginHash,
       context: context,
     );
   }
 
   void send() async {
     FocusScope.of(context).unfocus();
-    if(_ctrlStoreName.text.length == 0) {
+    if (_ctrlStoreName.text.length == 0) {
       Helpers.showAlertDialog(
-        context: context,
-        message: tr("store_name_empty"),
-        title: tr("action_error"),
-        onTapOk: null
-      );
+          context: context,
+          message: tr("store_name_empty"),
+          title: tr("action_error"),
+          onTapOk: null);
       return;
-    }
-    else if(selectedLatLng == null) {
+    } else if (selectedLatLng == null) {
       Helpers.showAlertDialog(
-        context: context,
-        title: tr("location_error"),
-        message: tr("location_error_details"),
-        onTapOk: null
-      );
+          context: context,
+          title: tr("location_error"),
+          message: tr("location_error_details"),
+          onTapOk: null);
       return;
-    }
-    else if(photoObjs.length == 0) {
+    } else if (photoObjs.length == 0) {
       Helpers.showAlertDialog(
-        context: context,
-        title: tr("photo_error"),
-        message: tr("photo_error_details"),
-        onTapOk: null
-      );
+          context: context,
+          title: tr("photo_error"),
+          message: tr("photo_error_details"),
+          onTapOk: null);
       return;
     }
 
     StoreServices.addStore(
-      storeId: widget.store.storeId, 
-      storeName: _ctrlStoreName.text, 
-      storeAddress: _ctrlStoreAddress.text, 
-      storeDesc: _ctrlStoreDesc.text, 
-      lat: selectedLatLng.latitude.toString(), 
-      lon: selectedLatLng.longitude.toString(), 
-      smsNo: _ctrlSMSNo.text, 
-      phoneNo: _ctrlPhoneNo.text, 
-      email: _ctrlEmail.text, 
-      website: _ctrlWebsite.text, 
-      categoryId: categories[selectedCategoryIndex].categoryId.toString(), 
-      monOpen: _timeStartMon, 
-      monClose: _timeEndMon, 
-      tueOpen: _timeStartTue, 
-      tueClose: _timeEndTue, 
-      wedOpen: _timeStartWed, 
-      wedClose: _timeEndWed,  
-      thuOpen: _timeStartThu, 
-      thuClose: _timeEndThu,
-      friOpen: _timeStartFri, 
-      friClose: _timeEndFri, 
-      satOpen: _timeStartSat, 
-      satClose: _timeEndSat, 
-      sunOpen: _timeStartSun, 
-      sunClose: _timeEndSun,
-      userId: loggedUser.userId, 
-      loginHash: loggedUser.loginHash,
-      photoPaths: photoObjs,
-      context: context,
-      photoIdsDeleted: photoIdsDeleted
-    );
+        storeId: widget.store?.storeId,
+        storeName: _ctrlStoreName.text,
+        storeAddress: _ctrlStoreAddress.text,
+        storeDesc: _ctrlStoreDesc.text,
+        lat: selectedLatLng!.latitude.toString(),
+        lon: selectedLatLng!.longitude.toString(),
+        smsNo: _ctrlSMSNo.text,
+        phoneNo: _ctrlPhoneNo.text,
+        email: _ctrlEmail.text,
+        website: _ctrlWebsite.text,
+        categoryId: categories![selectedCategoryIndex].categoryId.toString(),
+        monOpen: _timeStartMon,
+        monClose: _timeEndMon,
+        tueOpen: _timeStartTue,
+        tueClose: _timeEndTue,
+        wedOpen: _timeStartWed,
+        wedClose: _timeEndWed,
+        thuOpen: _timeStartThu,
+        thuClose: _timeEndThu,
+        friOpen: _timeStartFri,
+        friClose: _timeEndFri,
+        satOpen: _timeStartSat,
+        satClose: _timeEndSat,
+        sunOpen: _timeStartSun,
+        sunClose: _timeEndSun,
+        userId: loggedUser!.userId,
+        loginHash: loggedUser!.loginHash,
+        photoPaths: photoObjs,
+        context: context,
+        photoIdsDeleted: photoIdsDeleted);
   }
 
   List<Widget> _buildItems() {
-    return categories
+    return categories!
         .map((val) => CategorySelectionItem(
               category: val,
             ))
@@ -209,34 +201,31 @@ class _StoreEditScreenState extends State<StoreEditScreen> {
   }
 
   void setData() {
-
-    _ctrlStoreName..text = widget.store.storeName;
-    _ctrlStoreAddress..text = widget.store.storeAddress;
-    _ctrlStoreDesc..text = widget.store.storeDesc;
-    _ctrlSMSNo..text = widget.store.smsNo;
-    _ctrlPhoneNo..text = widget.store.phoneNo;
-    _ctrlEmail..text = widget.store.email;
-    _ctrlWebsite..text = widget.store.website;
+    _ctrlStoreName..text = widget.store!.storeName;
+    _ctrlStoreAddress..text = widget.store!.storeAddress;
+    _ctrlStoreDesc..text = widget.store!.storeDesc;
+    _ctrlSMSNo..text = widget.store!.smsNo;
+    _ctrlPhoneNo..text = widget.store!.phoneNo;
+    _ctrlEmail..text = widget.store!.email;
+    _ctrlWebsite..text = widget.store!.website;
 
     var lat = Config.DEFAULT_LAT;
     var lon = Config.DEFAULT_LON;
-    if(widget.store.lat != 0 && widget.store.lon != 0) {
-      lat = widget.store.lat;
-      lon = widget.store.lon;
+    if (widget.store!.lat != 0 && widget.store!.lon != 0) {
+      lat = widget.store!.lat;
+      lon = widget.store!.lon;
     }
-    
-    if(selectedLatLng == null)
-      selectedLatLng = LatLng(lat, lon);
 
-    if(selectedCategoryIndex == 0) {
-      for(int x = 0; x < categories.length; x++){
-        if(categories[x].categoryId == widget.store.categoryId) {
+    if (selectedLatLng == null) selectedLatLng = LatLng(lat, lon);
+
+    if (selectedCategoryIndex == 0) {
+      for (int x = 0; x < categories!.length; x++) {
+        if (categories![x].categoryId == widget.store!.categoryId) {
           selectedCategoryIndex = x;
           break;
         }
       }
     }
-
   }
 
   Widget showMain() {
@@ -247,31 +236,29 @@ class _StoreEditScreenState extends State<StoreEditScreen> {
         child: Scaffold(
           backgroundColor: Theme.of(context).backgroundColor,
           body: Column(
-            mainAxisAlignment: MainAxisAlignment.start,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: <Widget>[
-              getAppBarUI(),
-              main(),
-            ]
-          ),
+              mainAxisAlignment: MainAxisAlignment.start,
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: <Widget>[
+                getAppBarUI(),
+                main(),
+              ]),
         ),
       ),
     );
-    return root;
+    return root!;
   }
 
   @override
   Widget build(BuildContext context) {
     return FutureBuilder<bool>(
       future: getData(),
-      builder:(BuildContext context, AsyncSnapshot<bool> snapshot) {
+      builder: (BuildContext context, AsyncSnapshot<bool> snapshot) {
         if (!snapshot.hasData) {
           return LoadingWidget(
             size: 30.0,
             iconColor: Theme.of(context).accentColor,
           );
         } else {
-          
           return showMain();
         }
       },
@@ -280,45 +267,44 @@ class _StoreEditScreenState extends State<StoreEditScreen> {
 
   Widget main() {
     bool isDark = Helpers.isDarkMode(context);
-    String mapPlaceholder = isDark ? Config.DEFAULT_MAP_DARK_PLACEHOLDER : Config.DEFAULT_MAP_PLACEHOLDER;
-    String darkModeUrl = isDark ? Config.STATIC_DARK_GOOGLE_MAPS_URL : Config.STATIC_GOOGLE_MAPS_URL;
-    
+    String mapPlaceholder = isDark
+        ? Config.DEFAULT_MAP_DARK_PLACEHOLDER
+        : Config.DEFAULT_MAP_PLACEHOLDER;
+    String darkModeUrl = isDark
+        ? Config.STATIC_DARK_GOOGLE_MAPS_URL
+        : Config.STATIC_GOOGLE_MAPS_URL;
+
     var categoryWidget;
-    if(categories.length > 0) {
+    if (categories!.length > 0) {
       categoryWidget = DirectSelect(
-        itemExtent: 50.0,
-        mode: DirectSelectMode.tap,
-        selectedIndex: selectedCategoryIndex,
-        backgroundColor: Theme.of(context).accentColor,
-        child: CategorySelectionItem(
-          isForList: false,
-          category: categories[selectedCategoryIndex],
-        ),
-        onSelectedItemChanged: (index) {
-          
-          setState(() {
-            selectedCategoryIndex = index;
-          });
-        },
-        items: _buildItems()
-      );
-    }
-    else {
+          itemExtent: 50.0,
+          mode: DirectSelectMode.tap,
+          selectedIndex: selectedCategoryIndex,
+          backgroundColor: Theme.of(context).accentColor,
+          child: CategorySelectionItem(
+            isForList: false,
+            category: categories![selectedCategoryIndex],
+          ),
+          onSelectedItemChanged: (index) {
+            setState(() {
+              selectedCategoryIndex = index!;
+            });
+          },
+          items: _buildItems());
+    } else {
       categoryWidget = Helpers.createFormField(
-        controller: _ctrlStoreAddress,
-        placeholder: tr("no_categories_to_show"),
-        validatorText: null,
-        context: context,
-        enabled: false
-      );
+          controller: _ctrlStoreAddress,
+          placeholder: tr("no_categories_to_show"),
+          validatorText: null,
+          context: context,
+          enabled: false);
     }
 
     double leftPhoto = 0.0;
     double rightPhoto = 0.0;
-    if(Helpers.isRTL()) {
+    if (Helpers.isRTL()) {
       rightPhoto = 8.0;
-    }
-    else {
+    } else {
       leftPhoto = 8.0;
     }
 
@@ -331,296 +317,268 @@ class _StoreEditScreenState extends State<StoreEditScreen> {
               mainAxisAlignment: MainAxisAlignment.start,
               crossAxisAlignment: CrossAxisAlignment.start,
               children: <Widget>[
-
                 Helpers.createHeader(
                   context: context,
                   text: tr("store_name"),
                 ),
                 Helpers.createFormField(
-                  controller: _ctrlStoreName,
-                  placeholder: tr("store_name"),
-                  // validatorText: tr("store_name_empty"),
-                  context: context
-                ),
-
+                    controller: _ctrlStoreName,
+                    placeholder: tr("store_name"),
+                    // validatorText: tr("store_name_empty"),
+                    context: context),
                 Helpers.createHeader(
                   context: context,
                   text: tr("store_address"),
                 ),
                 Helpers.createFormField(
-                  controller: _ctrlStoreAddress,
-                  placeholder: tr("store_address"),
-                  validatorText: null,
-                  context: context
-                ),
-
+                    controller: _ctrlStoreAddress,
+                    placeholder: tr("store_address"),
+                    validatorText: null,
+                    context: context),
                 Helpers.createHeader(
                   context: context,
                   text: tr("sms_no"),
                 ),
                 Helpers.createFormField(
-                  controller: _ctrlSMSNo,
-                  placeholder: tr("sms_no"),
-                  validatorText: null,
-                  context: context
-                ),
-
+                    controller: _ctrlSMSNo,
+                    placeholder: tr("sms_no"),
+                    validatorText: null,
+                    context: context),
                 Helpers.createHeader(
                   context: context,
                   text: tr("phone_no"),
                 ),
                 Helpers.createFormField(
-                  controller: _ctrlPhoneNo,
-                  placeholder: tr("phone_no"),
-                  validatorText: null,
-                  context: context
-                ),
-
+                    controller: _ctrlPhoneNo,
+                    placeholder: tr("phone_no"),
+                    validatorText: null,
+                    context: context),
                 Helpers.createHeader(
                   context: context,
                   text: tr("email"),
                 ),
                 Helpers.createFormField(
-                  controller: _ctrlEmail,
-                  placeholder: tr("email"),
-                  validatorText: null,
-                  context: context
-                ),
-
+                    controller: _ctrlEmail,
+                    placeholder: tr("email"),
+                    validatorText: null,
+                    context: context),
                 Helpers.createHeader(
                   context: context,
                   text: tr("website"),
                 ),
                 Helpers.createFormField(
-                  controller: _ctrlWebsite,
-                  placeholder: tr("website"),
-                  validatorText: null,
-                  context: context
-                ),
-
+                    controller: _ctrlWebsite,
+                    placeholder: tr("website"),
+                    validatorText: null,
+                    context: context),
                 Helpers.createHeader(
                   context: context,
                   text: tr("category"),
                 ),
-                
                 categoryWidget,
-
                 Helpers.createHeader(
                   context: context,
                   text: tr("desc"),
                 ),
                 Helpers.createFormField(
-                  controller: _ctrlStoreDesc,
-                  placeholder: tr("desc"),
-                  validatorText: null,
-                  context: context,
-                  height: 150
+                    controller: _ctrlStoreDesc,
+                    placeholder: tr("desc"),
+                    validatorText: null,
+                    context: context,
+                    height: 150),
+                SizedBox(
+                  height: 50,
                 ),
-
-                SizedBox(height: 50,),
                 Helpers.createHeader(
                   context: context,
                   text: tr("operating_hours"),
                 ),
-
-                SizedBox(height: 20,),
+                SizedBox(
+                  height: 20,
+                ),
                 Helpers.createHeader(
                   context: context,
                   text: tr("mon"),
                 ),
                 Helpers.createTimeRange(
-                  context: context, 
-                  startTime: _timeStartMon, 
-                  endTime: _timeEndMon,
-                  onTimeStartChanged: (newTime) {
-                    setState(() {
-                      _timeStartMon = newTime;
-                    });
-                  },
-                  onTimeEndChanged: (newTime) {
-                    setState(() {
-                      _timeEndMon = newTime;
-                    });
-                  }
-                ),
-
+                    context: context,
+                    startTime: _timeStartMon,
+                    endTime: _timeEndMon,
+                    onTimeStartChanged: (newTime) {
+                      setState(() {
+                        _timeStartMon = newTime;
+                      });
+                    },
+                    onTimeEndChanged: (newTime) {
+                      setState(() {
+                        _timeEndMon = newTime;
+                      });
+                    }),
                 Helpers.createHeader(
                   context: context,
                   text: tr("tue"),
                 ),
                 Helpers.createTimeRange(
-                  context: context, 
-                  startTime: _timeStartTue, 
-                  endTime: _timeEndTue,
-                  onTimeStartChanged: (newTime) {
-                    setState(() {
-                      _timeStartTue = newTime;
-                    });
-                  },
-                  onTimeEndChanged: (newTime) {
-                    setState(() {
-                      _timeEndTue = newTime;
-                    });
-                  }
-                ),
-
+                    context: context,
+                    startTime: _timeStartTue,
+                    endTime: _timeEndTue,
+                    onTimeStartChanged: (newTime) {
+                      setState(() {
+                        _timeStartTue = newTime;
+                      });
+                    },
+                    onTimeEndChanged: (newTime) {
+                      setState(() {
+                        _timeEndTue = newTime;
+                      });
+                    }),
                 Helpers.createHeader(
                   context: context,
                   text: tr("wed"),
                 ),
                 Helpers.createTimeRange(
-                  context: context, 
-                  startTime: _timeStartWed, 
-                  endTime: _timeEndWed,
-                  onTimeStartChanged: (newTime) {
-                    setState(() {
-                      _timeStartWed = newTime;
-                    });
-                  },
-                  onTimeEndChanged: (newTime) {
-                    setState(() {
-                      _timeEndWed = newTime;
-                    });
-                  }
-                ),
-
+                    context: context,
+                    startTime: _timeStartWed,
+                    endTime: _timeEndWed,
+                    onTimeStartChanged: (newTime) {
+                      setState(() {
+                        _timeStartWed = newTime;
+                      });
+                    },
+                    onTimeEndChanged: (newTime) {
+                      setState(() {
+                        _timeEndWed = newTime;
+                      });
+                    }),
                 Helpers.createHeader(
                   context: context,
                   text: tr("thu"),
                 ),
                 Helpers.createTimeRange(
-                  context: context, 
-                  startTime: _timeStartThu, 
-                  endTime: _timeEndThu,
-                  onTimeStartChanged: (newTime) {
-                    setState(() {
-                      _timeStartThu = newTime;
-                    });
-                  },
-                  onTimeEndChanged: (newTime) {
-                    setState(() {
-                      _timeEndThu = newTime;
-                    });
-                  }
-                ),
-
+                    context: context,
+                    startTime: _timeStartThu,
+                    endTime: _timeEndThu,
+                    onTimeStartChanged: (newTime) {
+                      setState(() {
+                        _timeStartThu = newTime;
+                      });
+                    },
+                    onTimeEndChanged: (newTime) {
+                      setState(() {
+                        _timeEndThu = newTime;
+                      });
+                    }),
                 Helpers.createHeader(
                   context: context,
                   text: tr("fri"),
                 ),
                 Helpers.createTimeRange(
-                  context: context, 
-                  startTime: _timeStartFri, 
-                  endTime: _timeEndFri,
-                  onTimeStartChanged: (newTime) {
-                    setState(() {
-                      _timeStartFri = newTime;
-                    });
-                  },
-                  onTimeEndChanged: (newTime) {
-                    setState(() {
-                      _timeEndFri = newTime;
-                    });
-                  }
-                ),
-
+                    context: context,
+                    startTime: _timeStartFri,
+                    endTime: _timeEndFri,
+                    onTimeStartChanged: (newTime) {
+                      setState(() {
+                        _timeStartFri = newTime;
+                      });
+                    },
+                    onTimeEndChanged: (newTime) {
+                      setState(() {
+                        _timeEndFri = newTime;
+                      });
+                    }),
                 Helpers.createHeader(
                   context: context,
                   text: tr("sat"),
                 ),
                 Helpers.createTimeRange(
-                  context: context, 
-                  startTime: _timeStartSat, 
-                  endTime: _timeEndSat,
-                  onTimeStartChanged: (newTime) {
-                    setState(() {
-                      _timeStartSat = newTime;
-                    });
-                  },
-                  onTimeEndChanged: (newTime) {
-                    setState(() {
-                      _timeEndSat = newTime;
-                    });
-                  }
-                ),
-
+                    context: context,
+                    startTime: _timeStartSat,
+                    endTime: _timeEndSat,
+                    onTimeStartChanged: (newTime) {
+                      setState(() {
+                        _timeStartSat = newTime;
+                      });
+                    },
+                    onTimeEndChanged: (newTime) {
+                      setState(() {
+                        _timeEndSat = newTime;
+                      });
+                    }),
                 Helpers.createHeader(
                   context: context,
                   text: tr("sun"),
                 ),
                 Helpers.createTimeRange(
-                  context: context, 
-                  startTime: _timeStartSun, 
-                  endTime: _timeEndSun,
-                  onTimeStartChanged: (newTime) {
-                    setState(() {
-                      _timeStartSun = newTime;
-                    });
-                  },
-                  onTimeEndChanged: (newTime) {
-                    setState(() {
-                      _timeEndSun = newTime;
-                    });
-                  }
+                    context: context,
+                    startTime: _timeStartSun,
+                    endTime: _timeEndSun,
+                    onTimeStartChanged: (newTime) {
+                      setState(() {
+                        _timeStartSun = newTime;
+                      });
+                    },
+                    onTimeEndChanged: (newTime) {
+                      setState(() {
+                        _timeEndSun = newTime;
+                      });
+                    }),
+                SizedBox(
+                  height: 50,
                 ),
-
-                SizedBox(height: 50,),
                 Helpers.createHeader(
                   context: context,
                   text: tr("store_location"),
                 ),
-                SizedBox(height: 10,),
-                InkWell(
-                  onTap: () async{
-                    FocusScope.of(context).unfocus();
-                    DataHandler dataHandler = await Navigator.push<dynamic>(
-                      context,
-                      MaterialPageRoute<dynamic>(
-                          builder: (BuildContext context) => StoreMapScreen(
-                            lat: selectedLatLng.latitude,
-                            lon: selectedLatLng.longitude,
-                          ),
-                          fullscreenDialog: true),
-                    );
-
-                    if(dataHandler != null) {
-                      setState(() {
-                        imageBytes = dataHandler.bytes;
-                        selectedLatLng = dataHandler.latLng;
-                        
-                      });
-                    }
-                  },
-                  child: Container(
-                    height: 200,
-                    width: MediaQuery.of(context).size.width,
-                    child: ClipRRect(
-                      borderRadius: BorderRadius.circular(8),
-                      child: AspectRatio(
-                        aspectRatio: 1.2,
-                        child: imageBytes != null ?
-                          FadeInImage(
-                            fit: BoxFit.cover,
-                            placeholder: AssetImage(mapPlaceholder),
-                            image: MemoryImage(imageBytes),
-                          ) :
-                          FadeInImage(
-                            fit: BoxFit.cover,
-                            placeholder: AssetImage(mapPlaceholder),
-                            image: NetworkImage(sprintf(darkModeUrl, [
-                              selectedLatLng.longitude, 
-                              selectedLatLng.latitude, 
-                              selectedLatLng.longitude,
-                              selectedLatLng.latitude, 
-                              
-                            ])),
-                          )
-                      ),
-                    )
-                  )
+                SizedBox(
+                  height: 10,
                 ),
+                InkWell(
+                    onTap: () async {
+                      FocusScope.of(context).unfocus();
+                      DataHandler dataHandler = await Navigator.push<dynamic>(
+                        context,
+                        MaterialPageRoute<dynamic>(
+                            builder: (BuildContext context) => StoreMapScreen(
+                                  lat: selectedLatLng!.latitude,
+                                  lon: selectedLatLng!.longitude,
+                                ),
+                            fullscreenDialog: true),
+                      );
 
-                SizedBox(height: 50,),
+                      if (dataHandler != null) {
+                        setState(() {
+                          imageBytes = dataHandler.bytes;
+                          selectedLatLng = dataHandler.latLng;
+                        });
+                      }
+                    },
+                    child: Container(
+                        height: 200,
+                        width: MediaQuery.of(context).size.width,
+                        child: ClipRRect(
+                          borderRadius: BorderRadius.circular(8),
+                          child: AspectRatio(
+                              aspectRatio: 1.2,
+                              child: imageBytes != null
+                                  ? FadeInImage(
+                                      fit: BoxFit.cover,
+                                      placeholder: AssetImage(mapPlaceholder),
+                                      image: MemoryImage(imageBytes!),
+                                    )
+                                  : FadeInImage(
+                                      fit: BoxFit.cover,
+                                      placeholder: AssetImage(mapPlaceholder),
+                                      image: NetworkImage(sprintf(darkModeUrl, [
+                                        selectedLatLng?.longitude,
+                                        selectedLatLng?.latitude,
+                                        selectedLatLng?.longitude,
+                                        selectedLatLng?.latitude,
+                                      ])),
+                                    )),
+                        ))),
+                SizedBox(
+                  height: 50,
+                ),
                 Helpers.createHeader(
                   context: context,
                   text: tr("store_photos"),
@@ -634,60 +592,65 @@ class _StoreEditScreenState extends State<StoreEditScreen> {
                         Expanded(
                           child: Container(
                             decoration: BoxDecoration(
-                              color: Theme.of(context).textTheme.bodyText1.color,
+                              color:
+                                  Theme.of(context).textTheme.bodyText1?.color,
                               borderRadius: BorderRadius.circular(8),
                             ),
                             child: ListView.builder(
                               itemCount: photoObjs.length,
                               scrollDirection: Axis.horizontal,
                               itemBuilder: (BuildContext context, int index) {
-
                                 var widgetPhoto;
-                                NetworkImage widgetPhotoNetwork;
-                                if( photoObjs[index] is Photo ) {
+                                NetworkImage? widgetPhotoNetwork;
+                                if (photoObjs[index] is Photo) {
                                   Photo p = photoObjs[index];
                                   widgetPhotoNetwork = NetworkImage(p.photoUrl);
-                                }
-                                else {
-                                  widgetPhoto = FileImage(File(photoObjs[index]));
+                                } else {
+                                  widgetPhoto =
+                                      FileImage(File(photoObjs[index]));
                                 }
 
                                 return InkWell(
                                   child: Padding(
-                                    padding: EdgeInsets.only(left: leftPhoto, top: 8.0, bottom: 8.0, right: rightPhoto),
-                                    child: Container(
-                                      height: 140,
-                                      width: 140,
-                                      child: ClipRRect(
-                                        borderRadius: BorderRadius.circular(8),
-                                        child: AspectRatio(
-                                          aspectRatio: 1.2,
-                                          child: FadeInImage(
-                                            fit: BoxFit.cover,
-                                            placeholder: AssetImage(Helpers.getThemedPhotoPlaceHolder(context)),
-                                            image: widgetPhoto != null ? widgetPhoto : widgetPhotoNetwork,
-                                          )
-                                        ),
-                                      )
-                                    )
-                                  ),
+                                      padding: EdgeInsets.only(
+                                          left: leftPhoto,
+                                          top: 8.0,
+                                          bottom: 8.0,
+                                          right: rightPhoto),
+                                      child: Container(
+                                          height: 140,
+                                          width: 140,
+                                          child: ClipRRect(
+                                            borderRadius:
+                                                BorderRadius.circular(8),
+                                            child: AspectRatio(
+                                                aspectRatio: 1.2,
+                                                child: FadeInImage(
+                                                  fit: BoxFit.cover,
+                                                  placeholder: AssetImage(Helpers
+                                                      .getThemedPhotoPlaceHolder(
+                                                          context)),
+                                                  image: widgetPhoto != null
+                                                      ? widgetPhoto
+                                                      : widgetPhotoNetwork,
+                                                )),
+                                          ))),
                                   onTap: () {
                                     Helpers.showDualButtonDialog(
-                                      context: context,
-                                      title: tr("delete_photo"),
-                                      message: tr("delete_photo_details"),
-                                      onTapDismiss: () {
-
-                                      },
-                                      onTapOk: () {
-                                        if( photoObjs[index].runtimeType == Photo ) {
-                                          photoIdsDeleted.add(photoObjs[index].photoId);
-                                        }
-                                        setState(() {
-                                          photoObjs.removeAt(index);
+                                        context: context,
+                                        title: tr("delete_photo"),
+                                        message: tr("delete_photo_details"),
+                                        onTapDismiss: () {},
+                                        onTapOk: () {
+                                          if (photoObjs[index].runtimeType ==
+                                              Photo) {
+                                            photoIdsDeleted
+                                                .add(photoObjs[index].photoId);
+                                          }
+                                          setState(() {
+                                            photoObjs.removeAt(index);
+                                          });
                                         });
-                                      }
-                                    );
                                   },
                                 );
                               },
@@ -697,47 +660,51 @@ class _StoreEditScreenState extends State<StoreEditScreen> {
                         Padding(
                           padding: const EdgeInsets.only(left: 10),
                           child: Container(
-                            width: 50,
-                            height: MediaQuery.of(context).size.height,
-                            decoration: BoxDecoration(
-                              color: Theme.of(context).accentColor,
-                              borderRadius: BorderRadius.circular(8),
-                            ),
-                            child: Material(
-                              color: Theme.of(context).accentColor,
-                              borderRadius: BorderRadius.circular(8),
-                              child: InkWell(
+                              width: 50,
+                              height: MediaQuery.of(context).size.height,
+                              decoration: BoxDecoration(
+                                color: Theme.of(context).accentColor,
                                 borderRadius: BorderRadius.circular(8),
-                                child: Icon(
-                                  Icons.add_a_photo,
-                                  color: Theme.of(context).floatingActionButtonTheme.foregroundColor,
-                                ),
-                                onTap: () async {
-                                  if(photoObjs.length == Config.MAX_PHOTO_UPLOAD) {
-                                    Helpers.showAlertDialog(
-                                      context: context,
-                                      title: tr("max_photos_reached"),
-                                      message: tr("max_photos_reached_details"),
-                                    );
-                                    return;
-                                  }
-
-                                  PickedFile pickedFile = await picker.getImage(source: ImageSource.gallery, imageQuality: 90);
-                                  setState(() {
-                                    if (pickedFile != null) {
-                                      photoObjs.add(pickedFile.path);
-                                    }
-                                  });
-                                }
                               ),
-                            )
-                          ),
+                              child: Material(
+                                color: Theme.of(context).accentColor,
+                                borderRadius: BorderRadius.circular(8),
+                                child: InkWell(
+                                    borderRadius: BorderRadius.circular(8),
+                                    child: Icon(
+                                      Icons.add_a_photo,
+                                      color: Theme.of(context)
+                                          .floatingActionButtonTheme
+                                          .foregroundColor,
+                                    ),
+                                    onTap: () async {
+                                      if (photoObjs.length ==
+                                          Config.MAX_PHOTO_UPLOAD) {
+                                        Helpers.showAlertDialog(
+                                          context: context,
+                                          title: tr("max_photos_reached"),
+                                          message:
+                                              tr("max_photos_reached_details"),
+                                        );
+                                        return;
+                                      }
+
+                                      PickedFile? pickedFile =
+                                          await picker.getImage(
+                                              source: ImageSource.gallery,
+                                              imageQuality: 90);
+                                      setState(() {
+                                        if (pickedFile != null) {
+                                          photoObjs.add(pickedFile.path);
+                                        }
+                                      });
+                                    }),
+                              )),
                         )
                       ],
                     ),
                   ),
                 ),
-
                 Padding(
                   padding: const EdgeInsets.only(top: 16),
                   child: Center(
@@ -751,7 +718,9 @@ class _StoreEditScreenState extends State<StoreEditScreen> {
                         tr("update"),
                         style: TextStyle(
                           fontWeight: FontWeight.w500,
-                          color: Theme.of(context).floatingActionButtonTheme.foregroundColor,
+                          color: Theme.of(context)
+                              .floatingActionButtonTheme
+                              .foregroundColor,
                         ),
                       ),
                     ),
@@ -770,7 +739,9 @@ class _StoreEditScreenState extends State<StoreEditScreen> {
                         tr("delete"),
                         style: TextStyle(
                           fontWeight: FontWeight.w500,
-                          color: Theme.of(context).floatingActionButtonTheme.foregroundColor,
+                          color: Theme.of(context)
+                              .floatingActionButtonTheme
+                              .foregroundColor,
                         ),
                       ),
                     ),
@@ -787,7 +758,6 @@ class _StoreEditScreenState extends State<StoreEditScreen> {
     );
   }
 
-  
   Widget getAppBarUI() {
     return Container(
       decoration: BoxDecoration(
@@ -839,6 +809,4 @@ class _StoreEditScreenState extends State<StoreEditScreen> {
       ),
     );
   }
-  
 }
-
