@@ -1,13 +1,12 @@
-
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
 
 class LoadingWidget extends StatelessWidget {
   final double size;
-  final Color iconColor;
+  final Color? iconColor;
 
-  const LoadingWidget({Key key, this.size, this.iconColor})
+  const LoadingWidget({Key? key, this.size = 1.0, this.iconColor})
       : super(key: key);
 
   @override
