@@ -1,24 +1,26 @@
-
 import 'package:carousel_slider/carousel_slider.dart';
 import 'package:flutter/material.dart';
 import 'package:sprintf/sprintf.dart';
 import 'package:directorio_delicias/commons/helpers.dart';
 import 'package:directorio_delicias/models/store.dart';
 
-
 class SliderView extends StatefulWidget {
-  const SliderView({Key key, this.callBack, this.stores, this.onPageChanged}) : super(key: key);
+  const SliderView(
+      {Key? key,
+      required this.callBack,
+      required this.stores,
+      required this.onPageChanged})
+      : super(key: key);
 
   final Function(Store) callBack;
   final Function(int index, CarouselPageChangedReason reason) onPageChanged;
 
-  final List<Store> stores;
+  final List<Store>? stores;
   @override
   _SliderViewState createState() => _SliderViewState();
 }
 
 class _SliderViewState extends State<SliderView> {
-
   CarouselController controller = CarouselController();
 
   @override
@@ -44,19 +46,17 @@ class _SliderViewState extends State<SliderView> {
             return CarouselSlider(
               options: CarouselOptions(
                 height: 200.0,
-                onPageChanged: (index, reason) => widget.onPageChanged(index, reason),
+                onPageChanged: (index, reason) =>
+                    widget.onPageChanged(index, reason),
               ),
-              items: List.generate(
-                widget.stores.length, 
-                (int index) {
-                    return SliderWidget(
-                      callback: () {
-                        widget.callBack(widget.stores[index]);
-                      },
-                      store: widget.stores[index],
-                    );
-                }
-              ),
+              items: List.generate(widget.stores!.length, (int index) {
+                return SliderWidget(
+                  callback: () {
+                    widget.callBack(widget.stores![index]);
+                  },
+                  store: widget.stores![index],
+                );
+              }),
               carouselController: controller,
             );
           }
@@ -67,10 +67,7 @@ class _SliderViewState extends State<SliderView> {
 }
 
 class SliderWidget extends StatelessWidget {
-  const SliderWidget(
-      {Key key,
-      this.store,
-      this.callback})
+  const SliderWidget({Key? key, required this.store, required this.callback})
       : super(key: key);
 
   final VoidCallback callback;
@@ -79,9 +76,8 @@ class SliderWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     String photoUrl = "";
-    if(store.photos.length > 0)
-      photoUrl = store.photos[0].thumbUrl;
-      
+    if (store.photos!.length > 0) photoUrl = store.photos![0].thumbUrl;
+
     return InkWell(
       splashColor: Colors.transparent,
       onTap: () {
@@ -104,8 +100,9 @@ class SliderWidget extends StatelessWidget {
                       child: Container(
                         width: 290,
                         decoration: BoxDecoration(
-                          color: Theme.of(context).textTheme.bodyText2.color,
-                          borderRadius: const BorderRadius.all(Radius.circular(16.0)),
+                          color: Theme.of(context).textTheme.bodyText2?.color,
+                          borderRadius:
+                              const BorderRadius.all(Radius.circular(16.0)),
                         ),
                         child: Column(
                           children: <Widget>[
@@ -123,8 +120,10 @@ class SliderWidget extends StatelessWidget {
                                           right: 16,
                                           bottom: 8),
                                       child: Row(
-                                        mainAxisAlignment:MainAxisAlignment.spaceBetween,
-                                        crossAxisAlignment:CrossAxisAlignment.center,
+                                        mainAxisAlignment:
+                                            MainAxisAlignment.spaceBetween,
+                                        crossAxisAlignment:
+                                            CrossAxisAlignment.center,
                                         children: <Widget>[
                                           Expanded(
                                             child: Text(
@@ -136,7 +135,10 @@ class SliderWidget extends StatelessWidget {
                                                 fontWeight: FontWeight.w600,
                                                 fontSize: 16,
                                                 letterSpacing: 0.27,
-                                                color: Theme.of(context).textTheme.caption.color,
+                                                color: Theme.of(context)
+                                                    .textTheme
+                                                    .caption!
+                                                    .color,
                                               ),
                                             ),
                                           ),
@@ -144,18 +146,23 @@ class SliderWidget extends StatelessWidget {
                                             child: Row(
                                               children: <Widget>[
                                                 Text(
-                                                  sprintf("%.1f", [store.ratingAve]),
-                                                  textAlign:TextAlign.left,
+                                                  sprintf("%.1f",
+                                                      [store.ratingAve]),
+                                                  textAlign: TextAlign.left,
                                                   style: TextStyle(
-                                                    fontWeight:FontWeight.w200,
+                                                    fontWeight: FontWeight.w200,
                                                     fontSize: 18,
                                                     letterSpacing: 0.27,
-                                                    color: Theme.of(context).textTheme.subtitle1.color,
+                                                    color: Theme.of(context)
+                                                        .textTheme
+                                                        .subtitle1!
+                                                        .color,
                                                   ),
                                                 ),
                                                 Icon(
                                                   Icons.star,
-                                                  color:Theme.of(context).accentColor,
+                                                  color: Theme.of(context)
+                                                      .accentColor,
                                                   size: 20,
                                                 ),
                                               ],
@@ -168,23 +175,23 @@ class SliderWidget extends StatelessWidget {
                                 ),
                               ),
                             ),
-                            
                           ],
                         ),
                       ),
                     ),
-                    
                   ],
                 ),
               ),
               Container(
                 child: Padding(
-                  padding: const EdgeInsets.only(top: 16, right: 16, left: 16, bottom: 16),
+                  padding: const EdgeInsets.only(
+                      top: 16, right: 16, left: 16, bottom: 16),
                   child: Container(
                     width: 280,
                     height: 140,
                     decoration: BoxDecoration(
-                      borderRadius: const BorderRadius.all(Radius.circular(16.0)),
+                      borderRadius:
+                          const BorderRadius.all(Radius.circular(16.0)),
                       boxShadow: <BoxShadow>[
                         BoxShadow(
                             color: Theme.of(context).shadowColor,
@@ -193,14 +200,14 @@ class SliderWidget extends StatelessWidget {
                       ],
                     ),
                     child: ClipRRect(
-                      borderRadius: const BorderRadius.all(Radius.circular(16.0)),
+                      borderRadius:
+                          const BorderRadius.all(Radius.circular(16.0)),
                       child: AspectRatio(
                           aspectRatio: 1.28,
                           child: AspectRatio(
                             aspectRatio: 1.0,
                             child: Helpers.loadCacheImage(imageUrl: photoUrl),
-                          )
-                        ),
+                          )),
                     ),
                   ),
                 ),
