@@ -1,7 +1,6 @@
-
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
-import 'package:smooth_star_rating/smooth_star_rating.dart';
+import 'package:smooth_star_rating_nsafe/smooth_star_rating.dart';
 import 'package:sprintf/sprintf.dart';
 import 'package:directorio_delicias/commons/helpers.dart';
 import 'package:directorio_delicias/db/db_provider.dart';
@@ -10,27 +9,25 @@ import 'package:directorio_delicias/models/store.dart';
 
 class StoreListView extends StatelessWidget {
   const StoreListView(
-      {Key key,
-      this.store,
-      this.callback,
+      {Key? key,
+      required this.store,
+      required this.callback,
       this.faveCallback,
       this.isFave = false})
       : super(key: key);
 
   final VoidCallback callback;
-  final Function(bool) faveCallback;
+  final Function(bool)? faveCallback;
   final Store store;
   final bool isFave;
 
   @override
   Widget build(BuildContext context) {
     String photoUrl = "";
-    if(store.photos.length > 0)
-      photoUrl = store.photos[0].thumbUrl;
-      
+    if (store.photos!.length > 0) photoUrl = store.photos![0].thumbUrl;
+
     return Padding(
-      padding: const EdgeInsets.only(
-          left: 24, right: 24, top: 8, bottom: 16),
+      padding: const EdgeInsets.only(left: 24, right: 24, top: 8, bottom: 16),
       child: InkWell(
         splashColor: Colors.transparent,
         onTap: () {
@@ -58,7 +55,7 @@ class StoreListView extends StatelessWidget {
                       child: Helpers.loadCacheImage(imageUrl: photoUrl),
                     ),
                     Container(
-                      color: Theme.of(context).textTheme.bodyText1.color,
+                      color: Theme.of(context).textTheme.bodyText1?.color,
                       child: Row(
                         mainAxisAlignment: MainAxisAlignment.center,
                         crossAxisAlignment: CrossAxisAlignment.start,
@@ -69,8 +66,8 @@ class StoreListView extends StatelessWidget {
                                 padding: const EdgeInsets.only(
                                     left: 16, top: 8, bottom: 8, right: 16),
                                 child: Column(
-                                  mainAxisAlignment:MainAxisAlignment.center,
-                                  crossAxisAlignment:CrossAxisAlignment.start,
+                                  mainAxisAlignment: MainAxisAlignment.center,
+                                  crossAxisAlignment: CrossAxisAlignment.start,
                                   children: <Widget>[
                                     Text(
                                       store.storeName,
@@ -80,29 +77,34 @@ class StoreListView extends StatelessWidget {
                                       style: TextStyle(
                                         fontWeight: FontWeight.w600,
                                         fontSize: 22,
-                                        color: Theme.of(context).textTheme.caption.color,
+                                        color: Theme.of(context)
+                                            .textTheme
+                                            .caption!
+                                            .color,
                                       ),
                                     ),
                                     Row(
-                                      crossAxisAlignment:CrossAxisAlignment.center,
-                                      mainAxisAlignment:MainAxisAlignment.start,
+                                      crossAxisAlignment:
+                                          CrossAxisAlignment.center,
+                                      mainAxisAlignment:
+                                          MainAxisAlignment.start,
                                       children: <Widget>[
                                         Expanded(
-                                            child: Text(
-                                            store.storeAddress,
-                                            overflow: TextOverflow.ellipsis,
-                                            maxLines: 2,
-                                            style: TextStyle(
+                                          child: Text(store.storeAddress,
+                                              overflow: TextOverflow.ellipsis,
+                                              maxLines: 2,
+                                              style: TextStyle(
                                                 fontSize: 14,
-                                                color: Theme.of(context).textTheme.subtitle1.color,
-                                            )
-                                          ),
+                                                color: Theme.of(context)
+                                                    .textTheme
+                                                    .subtitle1!
+                                                    .color,
+                                              )),
                                         )
                                       ],
                                     ),
                                     Padding(
-                                      padding:
-                                          const EdgeInsets.only(top: 4),
+                                      padding: const EdgeInsets.only(top: 4),
                                       child: Row(
                                         children: <Widget>[
                                           SmoothStarRating(
@@ -110,16 +112,22 @@ class StoreListView extends StatelessWidget {
                                             starCount: 5,
                                             rating: store.ratingAve,
                                             size: 20,
-                                            color: Theme.of(context).accentColor,
-                                            borderColor: Theme.of(context).accentColor,
+                                            color:
+                                                Theme.of(context).accentColor,
+                                            borderColor:
+                                                Theme.of(context).accentColor,
                                           ),
                                           Text(
-                                            sprintf("%.1f %s", 
-                                              [store.ratingAve,
-                                              tr("ratings")]),
+                                            sprintf("%.1f %s", [
+                                              store.ratingAve,
+                                              tr("ratings")
+                                            ]),
                                             style: TextStyle(
                                                 fontSize: 14,
-                                                color: Theme.of(context).textTheme.subtitle1.color),
+                                                color: Theme.of(context)
+                                                    .textTheme
+                                                    .subtitle1!
+                                                    .color),
                                           ),
                                         ],
                                       ),
@@ -131,22 +139,20 @@ class StoreListView extends StatelessWidget {
                           ),
                           Padding(
                             padding: const EdgeInsets.only(
-                                right: 16, top: 8, left:16),
+                                right: 16, top: 8, left: 16),
                             child: Column(
-                              mainAxisAlignment:
-                                  MainAxisAlignment.center,
-                              crossAxisAlignment:
-                                  CrossAxisAlignment.end,
+                              mainAxisAlignment: MainAxisAlignment.center,
+                              crossAxisAlignment: CrossAxisAlignment.end,
                               children: <Widget>[
                                 Text(
-                                  sprintf("%.1f %s", 
-                                    [store.distance,
-                                    tr("km")]
-                                  ),
+                                  sprintf(
+                                      "%.1f %s", [store.distance, tr("km")]),
                                   style: TextStyle(
-                                    fontSize: 14,
-                                    color: Theme.of(context).textTheme.subtitle1.color
-                                  ),
+                                      fontSize: 14,
+                                      color: Theme.of(context)
+                                          .textTheme
+                                          .subtitle1!
+                                          .color),
                                 ),
                               ],
                             ),
@@ -161,29 +167,29 @@ class StoreListView extends StatelessWidget {
                   right: 8,
                   child: Material(
                     color: Colors.transparent,
-                    child: FutureBuilder<Fave>(
-                      future: DBProvider.instance.getFaveByStoreId(store.storeId),
-                      builder: (BuildContext context, AsyncSnapshot<Fave> snapshot) {
-                        if(!snapshot.hasData || snapshot == null) {
-                          return Padding(
-                            padding: const EdgeInsets.all(8.0),
-                            child: Icon(
-                              Icons.favorite_border,
-                              color: Theme.of(context).accentColor,
-                            ),
-                          );
-                        }
-                        else {
-                          return Padding(
-                            padding: const EdgeInsets.all(8.0),
-                            child: Icon(
-                              Icons.favorite,
-                              color: Theme.of(context).accentColor,
-                            ),
-                          );
-                        }
-                      }
-                    ),
+                    child: FutureBuilder<Fave?>(
+                        future:
+                            DBProvider.instance.getFaveByStoreId(store.storeId),
+                        builder: (BuildContext context,
+                            AsyncSnapshot<Fave?> snapshot) {
+                          if (!snapshot.hasData || snapshot == null) {
+                            return Padding(
+                              padding: const EdgeInsets.all(8.0),
+                              child: Icon(
+                                Icons.favorite_border,
+                                color: Theme.of(context).accentColor,
+                              ),
+                            );
+                          } else {
+                            return Padding(
+                              padding: const EdgeInsets.all(8.0),
+                              child: Icon(
+                                Icons.favorite,
+                                color: Theme.of(context).accentColor,
+                              ),
+                            );
+                          }
+                        }),
                   ),
                 )
               ],
