@@ -1,6 +1,5 @@
-
 import 'package:flutter_login_facebook/flutter_login_facebook.dart';
-import 'package:flutter_twitter/flutter_twitter.dart';
+
 import 'package:get_it/get_it.dart';
 import 'package:sprintf/sprintf.dart';
 import 'package:directorio_delicias/app/config.dart';
@@ -8,42 +7,42 @@ import 'package:directorio_delicias/dataparser/data_handler.dart';
 import 'package:directorio_delicias/dataparser/data_parser.dart';
 import 'package:directorio_delicias/db/db_provider.dart';
 import 'package:directorio_delicias/main.dart';
+import 'package:twitter_login/twitter_login.dart';
 
 class AuthFetcher {
-  
   Function(DataHandler) didLogged;
   Function(DataHandler) didError;
-  AuthFetcher({this.didLogged, this.didError});
+  AuthFetcher({required this.didLogged, required this.didError});
 
   Future<void> loginTwitter() async {
-    final twitterLogin = TwitterLogin(  
-      consumerKey: Config.TWITTER_CONSUMER_KEY,
-      consumerSecret: Config.TWITTER_CONSUMER_SECRET,
+    final twitterLogin = TwitterLogin(
+      redirectURI: "",
+      apiKey: Config.TWITTER_CONSUMER_KEY,
+      apiSecretKey: Config.TWITTER_CONSUMER_SECRET,
     );
-    
-    TwitterLoginResult authResult = await twitterLogin.authorize();
-    if(authResult.status == TwitterLoginStatus.loggedIn) {
-      String profileImage = sprintf("https://twivatar.glitch.me/%s", [authResult.session.username]);
-      DataHandler dataHandler = await GetIt.instance.get<DataParser>().register(
-        "", 
-        "", 
-        "",
-        authResult.session.username,
-        "",
-        authResult.session.userId,
-        "",
-        profileImage,
-        null
-      );
 
-      if(dataHandler.user != null) {
+    final authResult = await twitterLogin.login();
+    if (authResult.status == TwitterLoginStatus.loggedIn) {
+      String profileImage =
+          sprintf("https://twivatar.glitch.me/%s", [authResult.user?.name]);
+      DataHandler dataHandler = await GetIt.instance.get<DataParser>().register(
+          "",
+          "",
+          "",
+          authResult.user?.screenName,
+          "",
+          authResult.user?.id,
+          "",
+          profileImage,
+          null);
+
+      if (dataHandler.user != null) {
         DBProvider.instance.deleteLoggedUser();
-        DBProvider.instance.insertLoggedUser(dataHandler.user);
+        DBProvider.instance.insertLoggedUser(dataHandler.user!);
         MyApp.loggedUser = dataHandler.user;
-        if(this.didLogged != null) this.didLogged(dataHandler);
-      }
-      else {
-        if(this.didError != null) this.didError(dataHandler);
+        if (this.didLogged != null) this.didLogged(dataHandler);
+      } else {
+        if (this.didError != null) this.didError(dataHandler);
       }
     }
   }
@@ -57,10 +56,10 @@ class AuthFetcher {
 
     // Check result status
     switch (res.status) {
-      case FacebookLoginStatus.Success:
+      case FacebookLoginStatus.success:
         // Logged in
         // Send access token to server for validation and auth
-        final FacebookAccessToken accessToken = res.accessToken;
+        final FacebookAccessToken? accessToken = res.accessToken;
         print(accessToken);
         // Get profile data
         final profile = await fb.getUserProfile();
@@ -71,92 +70,78 @@ class AuthFetcher {
         final email = await fb.getUserEmail();
         // But user can decline permission
 
-        DataHandler dataHandler = await GetIt.instance.get<DataParser>().register(
-          "", 
-          "", 
-          email == null ? "" : email,
-          profile.name,
-          profile.userId,
-          "",
-          "",
-          imageUrl,
-          null
-        );
+        DataHandler dataHandler = await GetIt.instance
+            .get<DataParser>()
+            .register("", "", email == null ? "" : email, profile?.name,
+                profile?.userId, "", "", imageUrl, null);
 
-        if(dataHandler.user != null) {
+        if (dataHandler.user != null) {
           DBProvider.instance.deleteLoggedUser();
-          DBProvider.instance.insertLoggedUser(dataHandler.user);
+          DBProvider.instance.insertLoggedUser(dataHandler.user!);
           MyApp.loggedUser = dataHandler.user;
-          if(this.didLogged != null) this.didLogged(dataHandler);
-        }
-        else {
-          if(this.didError != null) this.didError(dataHandler);
+          if (this.didLogged != null) this.didLogged(dataHandler);
+        } else {
+          if (this.didError != null) this.didError(dataHandler);
         }
         break;
-      case FacebookLoginStatus.Cancel:
+      case FacebookLoginStatus.cancel:
         // User cancel log in
         break;
-      case FacebookLoginStatus.Error:
+      case FacebookLoginStatus.error:
         // Log in failed
         print('Error while log in: ${res.error}');
         break;
+      case FacebookLoginStatus.success:
+        // TODO: Handle this case.
+        break;
+      case FacebookLoginStatus.cancel:
+        // TODO: Handle this case.
+        break;
+      case FacebookLoginStatus.error:
+        // TODO: Handle this case.
+        break;
     }
-  } 
+  }
 
   void loginUser({username, password}) async {
-    DataHandler dataHandler = await GetIt.instance.get<DataParser>().login(username, password);
-    if(dataHandler.user != null) {
-      DBProvider.instance.insertLoggedUser(dataHandler.user);
+    DataHandler dataHandler =
+        await GetIt.instance.get<DataParser>().login(username, password);
+    if (dataHandler.user != null) {
+      DBProvider.instance.insertLoggedUser(dataHandler.user!);
       MyApp.loggedUser = dataHandler.user;
-      if(this.didLogged != null) this.didLogged(dataHandler);
+      if (this.didLogged != null) this.didLogged(dataHandler);
+    } else {
+      if (this.didLogged != null) this.didError(dataHandler);
     }
-    else {
-      if(this.didLogged != null) this.didError(dataHandler);
-    }
-  } 
+  }
 
   void register({username, password, email, fullName, photoFile}) async {
     DataHandler dataHandler = await GetIt.instance.get<DataParser>().register(
-      username, 
-      password, 
-      email, 
-      fullName, 
-      "",
-      "",
-      "",
-      "",
-      photoFile
-    );
+        username, password, email, fullName, "", "", "", "", photoFile);
 
-    if(dataHandler.user != null) {
-      DBProvider.instance.insertLoggedUser(dataHandler.user);
+    if (dataHandler.user != null) {
+      DBProvider.instance.insertLoggedUser(dataHandler.user!);
       MyApp.loggedUser = dataHandler.user;
-      if(this.didLogged != null) this.didLogged(dataHandler);
+      if (this.didLogged != null) this.didLogged(dataHandler);
+    } else {
+      if (this.didLogged != null) this.didError(dataHandler);
     }
-    else {
-      if(this.didLogged != null) this.didError(dataHandler);
-    }
-  } 
+  }
 
   void updateProfile({userId, password, fullName, thumbUrl, photoFile}) async {
-    DataHandler dataHandler = await GetIt.instance.get<DataParser>().updateProfile(
-      userId, 
-      password, 
-      fullName, 
-      thumbUrl,
-      photoFile
-    );
+    DataHandler dataHandler = await GetIt.instance
+        .get<DataParser>()
+        .updateProfile(userId, password, fullName, thumbUrl, photoFile);
 
-    if(dataHandler.user != null) {
+    if (dataHandler.user != null) {
       DBProvider.instance.deleteLoggedUser();
-      DBProvider.instance.insertLoggedUser(dataHandler.user);
+      DBProvider.instance.insertLoggedUser(dataHandler.user!);
       MyApp.loggedUser = dataHandler.user;
-      if(this.didLogged != null) this.didLogged(dataHandler);
+      if (this.didLogged != null) this.didLogged(dataHandler);
+    } else {
+      if (this.didLogged != null) this.didError(dataHandler);
     }
-    else {
-      if(this.didLogged != null) this.didError(dataHandler);
-    }
-  } 
+  }
 
   void logoutUser() {
     FacebookLogin fbController = FacebookLogin(debug: true);
@@ -165,6 +150,6 @@ class AuthFetcher {
     MyApp.loggedUser = null;
     DataHandler dataHandler = new DataHandler();
     dataHandler.isLoggedOut = true;
-    if(this.didLogged != null) this.didLogged(dataHandler);
+    if (this.didLogged != null) this.didLogged(dataHandler);
   }
 }
